Split continues — start, metadata, prefs, extract tabs moved to core/tabs/ .
Continues the monolith split. Four GUI tabs are relocated into
fizgig_gui/core/tabs/{start,metadata,prefs,extract}_tab.py. Pure move: each
tab class keeps its methods and dependencies. Monolith: 25,101 → 22,474 lines.
Per-module refactoring is deferred.
