+++ lora_trainer_gui.py
@@ -97,6 +97,10 @@ from fizgig_gui.core.config.presets import (
 )
 from fizgig_gui.core.ui_base.styling import StylingMixin
 from fizgig_gui.core.ui_base.tab_scaffold import TabScaffoldMixin
+from fizgig_gui.core.tabs.start_tab import StartTabMixin
+from fizgig_gui.core.tabs.metadata_tab import MetadataTabMixin
+from fizgig_gui.core.tabs.prefs_tab import PrefsTabMixin
+from fizgig_gui.core.tabs.extract_tab import ExtractTabMixin
 from fizgig_gui.core.ui_base.console_validation import ConsoleValidationMixin
 
 from fizgig_gui.core.config.last_used import (
@@ -120,7 +124,11 @@ from fizgig_gui.core.config.prefs import (
 
 from fizgig_gui.core.config.settings_map import SETTING_TO_PREF, PRESETS
 
-class LoRATrainerGUI(StylingMixin, TabScaffoldMixin, ConsoleValidationMixin):
+class LoRATrainerGUI(
+    StartTabMixin, MetadataTabMixin, PrefsTabMixin,
+    ExtractTabMixin, StylingMixin, TabScaffoldMixin,
+    ConsoleValidationMixin
+):
     def __init__(self, master):
         self.master = master
         master.title("Fizgig — Klein 9B & Krea 2 LoRA Studio")
@@ -1419,425 +1427,6 @@ class LoRATrainerGUI(StylingMixin, TabScaffoldMixin, ConsoleValidationMixin):
             initialdir=self._pref_initialdir("input_ref_dir"))
         if path:
             self.sample_ref_image_var.set(path)
-
-    def create_start_tab(self):
-        """Welcome screen + Training image folder picker — the single source of truth shared with
-        the Image Prep / Captions tabs and the Fizgig_train.toml auto-saver."""
-        scrollable_frame, _ = self.create_scrollable_frame(self.start_tab)
-
-        container = tk.Frame(scrollable_frame, bg=COLORS["bg_deep"])
-        container.pack(fill=tk.BOTH, expand=True, padx=36, pady=(28, 0))
-
-        # Title + subtitle
-        tk.Label(container, text="Welcome to Fizgig",
-                 font=(FONT_FAMILY, 22, "bold"),
-                 fg=COLORS["text_primary"], bg=COLORS["bg_deep"]).pack(anchor=tk.W)
-        tk.Label(container,
-                 text="A focused, local trainer and workbench for Flux 2 Klein 9B, Krea 2 and "
-                      "MiniMax H3 LoRAs — train, profile, repair, explore, and extract, all in "
-                      "one place.",
-                 font=(FONT_FAMILY, 11),
-                 fg=COLORS["text_secondary"], bg=COLORS["bg_deep"],
-                 wraplength=800, justify=tk.LEFT).pack(anchor=tk.W, pady=(4, 24))
-
-        # Workflow card
-        workflow_card = tk.Frame(container, bg=COLORS["bg_surface"],
-                                 highlightbackground=COLORS["border"],
-                                 highlightthickness=1, bd=0)
-        workflow_card.pack(fill=tk.X, pady=(0, 20))
-
-        # Use grid on workflow_card: col 0 = title+steps, col 1 = logo (no padding)
-        workflow_card.columnconfigure(0, weight=1)
-
-        tk.Label(workflow_card, text="Training Workflow",
-                 font=(FONT_FAMILY, 12, "bold"),
-                 fg=COLORS["text_primary"], bg=COLORS["bg_surface"]).grid(
-            row=0, column=0, sticky=tk.W, padx=20, pady=(16, 10))
-
-        steps_frame = tk.Frame(workflow_card, bg=COLORS["bg_surface"])
-        steps_frame.grid(row=1, column=0, sticky=tk.NSEW, padx=20, pady=(0, 16))
-
-        # Logo — spans both rows, fills full card height, zero padding
-        try:
-            from PIL import Image as _PILImage, ImageTk as _PILImageTk
-            _logo_path = os.path.join(os.path.dirname(__file__), "logo.jpg")
-            if os.path.exists(_logo_path):
-                _logo_pil_src = _PILImage.open(_logo_path)
-                _logo_label = tk.Label(workflow_card, bg=COLORS["bg_surface"])
-                _logo_label.grid(row=0, column=1, rowspan=2, sticky=tk.NS, padx=0, pady=0)
-
-                def _fit_logo(_event=None, _src=_logo_pil_src, _lbl=_logo_label):
-                    h = workflow_card.winfo_height()
-                    if h < 20:
-                        h = 260
-                    w = int(_src.width * h / _src.height)
-                    resized = _src.resize((w, h), _PILImage.LANCZOS)
-                    self._start_logo_tk = _PILImageTk.PhotoImage(resized)
-                    _lbl.configure(image=self._start_logo_tk)
-
-                self.master.after(100, _fit_logo)
-        except Exception:
-            pass  # PIL not available or logo missing — skip silently
-
-        steps = [
-            ("1", "Start",      "Choose your training image folder below.",                     False),
-            ("2", "Image Prep", "Resize, convert to PNG, or face-crop. (Video Prep for MiniMax)", True),  # optional
-            ("3", "Captions",   "Write trigger-word captions or generate them with AI.",        False),
-            ("4", "Samples",    "Configure in-training preview prompts.",                       False),
-            ("5", "Training",   "Pick a preset, tune settings, click Start Training.",          False),
-        ]
-
-        for num, tab_name, desc, is_optional in steps:
-            row = tk.Frame(steps_frame, bg=COLORS["bg_surface"])
-            row.pack(fill=tk.X, pady=4)
-
-            # Step number badge
-            tk.Label(row, text=num,
-                     font=(FONT_FAMILY, 11, "bold"),
-                     fg=COLORS["accent"], bg=COLORS["bg_surface"],
-                     width=2).pack(side=tk.LEFT, padx=(0, 12))
-
-            # Tab name
-            tk.Label(row, text=tab_name,
-                     font=(FONT_FAMILY, 11, "bold"),
-                     fg=COLORS["text_primary"], bg=COLORS["bg_surface"],
-                     width=12, anchor="w").pack(side=tk.LEFT)
-
-            # Optional badge
-            if is_optional:
-                tk.Label(row, text="OPTIONAL",
-                         font=(FONT_FAMILY, 8, "bold"),
-                         fg=COLORS["warning"], bg=COLORS["bg_surface"]).pack(side=tk.LEFT, padx=(0, 10))
-
-            # Description
-            tk.Label(row, text=desc,
-                     font=(FONT_FAMILY, 10),
-                     fg=COLORS["text_secondary"], bg=COLORS["bg_surface"],
-                     anchor="w", justify=tk.LEFT).pack(side=tk.LEFT, fill=tk.X, expand=True)
-
-        # Folder picker card
-        picker_card = tk.Frame(container, bg=COLORS["bg_surface"],
-                               highlightbackground=COLORS["accent"],
-                               highlightthickness=1, bd=0)
-        picker_card.pack(fill=tk.X, pady=(0, 4))
-
-        tk.Label(picker_card, text="Training image folder",
-                 font=(FONT_FAMILY, 12, "bold"),
-                 fg=COLORS["text_primary"], bg=COLORS["bg_surface"]).pack(anchor=tk.W, padx=20, pady=(8, 2))
-        tk.Label(picker_card,
-                 text="This is the single place you set your dataset folder. Image Prep, Captions, "
-                      "and Training all read from it automatically.",
-                 font=(FONT_FAMILY, 10),
-                 fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                 wraplength=760, justify=tk.LEFT).pack(anchor=tk.W, padx=20, pady=(0, 12))
-
-        row = tk.Frame(picker_card, bg=COLORS["bg_surface"])
-        row.pack(fill=tk.X, padx=20, pady=(0, 8))
-        ttk.Entry(row, textvariable=self.image_folder_var, width=70, font=(FONT_FAMILY, 10)).pack(
-            side=tk.LEFT, fill=tk.X, expand=True, padx=(0, 10), ipady=4
-        )
-        ttk.Button(row, text="Browse…", command=self._browse_image_folder).pack(side=tk.LEFT)
-
-        # Setup prompt — shown when model paths are not configured yet
-        self._setup_prompt_frame = tk.Frame(container, bg=COLORS["warning"],
-                                             highlightbackground=COLORS["warning"],
-                                             highlightthickness=1, bd=0)
-        setup_inner = tk.Frame(self._setup_prompt_frame, bg="#2A2200")
-        setup_inner.pack(fill=tk.X, padx=1, pady=1)
-        tk.Label(setup_inner, text="\u26a0  Model files not configured",
-                 font=(FONT_FAMILY, 12, "bold"),
-                 fg=COLORS["warning"], bg="#2A2200").pack(anchor=tk.W, padx=20, pady=(12, 4))
-        tk.Label(setup_inner,
-                 text="Head to the Preferences tab to set your model paths before training or using the tools. "
-                      "Each model row has a Download link that opens the correct HuggingFace page.",
-                 font=(FONT_FAMILY, 10),
-                 fg=COLORS["text_primary"], bg="#2A2200",
-                 wraplength=760, justify=tk.LEFT).pack(anchor=tk.W, padx=20, pady=(0, 4))
-        _setup_btn_row = tk.Frame(setup_inner, bg="#2A2200")
-        _setup_btn_row.pack(anchor=tk.W, padx=20, pady=(4, 12))
-        ttk.Button(_setup_btn_row, text="Open Preferences",
-                   command=lambda: self.notebook.select(self.prefs_tab)).pack(side=tk.LEFT)
-
-        def _dismiss_setup_prompt():
-            # Permanent, by request: a Krea-only user never fills the Klein paths (or vice
-            # versa, or skips the Turbo checkpoint entirely) and shouldn't be nagged forever.
-            self.prefs["setup_prompt_dismissed"] = True
-            save_prefs(self.prefs)
-            self._setup_prompt_frame.pack_forget()
-
-        _dismiss_lbl = tk.Label(_setup_btn_row, text="Don't show this again",
-                                font=(FONT_FAMILY, 9, "underline"),
-                                fg=COLORS["text_secondary"], bg="#2A2200", cursor="hand2")
-        _dismiss_lbl.pack(side=tk.LEFT, padx=(16, 0))
-        _dismiss_lbl.bind("<Button-1>", lambda e: _dismiss_setup_prompt())
-
-        def _check_model_paths(*_args):
-            # Hidden forever once dismissed; otherwise satisfied by EITHER family being
-            # usable — Klein's four paths, or Krea 2's training trio (the Turbo checkpoint
-            # is optional now that previews default to the Turbo LoRA).
-            if self.prefs.get("setup_prompt_dismissed"):
-                self._setup_prompt_frame.pack_forget()
-                return
-            klein_ok = all(self.prefs_vars[k].get().strip()
-                           for k in ("base_dit", "distilled_dit", "vae", "text_encoder"))
-            krea_ok = all(self.prefs_vars[k].get().strip()
-                          for k in ("krea2_raw_dit", "krea2_vae", "krea2_text_encoder"))
-            if klein_ok or krea_ok:
-                self._setup_prompt_frame.pack_forget()
-            else:
-                self._setup_prompt_frame.pack(fill=tk.X, pady=(20, 0),
-                                               before=tools_card)
-
-        # Re-check whenever a model path (either family) changes
-        for _mk in ("base_dit", "distilled_dit", "vae", "text_encoder",
-                    "krea2_raw_dit", "krea2_vae", "krea2_text_encoder"):
-            self.prefs_vars[_mk].trace_add("write", _check_model_paths)
-
-        # Initial check (deferred so tools_card exists)
-        self.master.after(100, _check_model_paths)
-
-        def _check_minimax_extras():
-            # An H3 user (DiT set) missing the NEW files — the Audio VAE and the Turbo LoRA
-            # both arrived after most people set up their paths, so nothing else would ever
-            # tell them these exist (Peter). One popup, dismissable forever.
-            if self.prefs.get("minimax_extras_prompt_dismissed"):
-                return
-            if not str(self.prefs.get("minimax_dit", "") or "").strip():
-                return
-            missing = [label for key, label in (
-                ("minimax_audio_vae",
-                 "Audio VAE (~605 MB) — train on the sound in video clips, and on voices"),
-                ("minimax_turbo_lora",
-                 "Turbo LoRA (~780 MB) — fast 6-step in-training previews"))
-                if not str(self.prefs.get(key, "") or "").strip()]
-            if not missing:
-                return
-            win = tk.Toplevel(self.master)
-            win.title("MiniMax H3 — new model files")
-            win.configure(bg=COLORS["bg_deep"], padx=20, pady=16)
-            win.transient(self.master)
-            win.resizable(False, False)
-            tk.Label(win, text="Your MiniMax H3 setup is missing the new files",
-                     font=(FONT_FAMILY, 12, "bold"), bg=COLORS["bg_deep"],
-                     fg=COLORS["text_primary"]).pack(anchor=tk.W)
-            tk.Label(win, text="Fizgig can now train on video, sound and voices, and render "
-                               "fast Turbo previews. Your H3 model paths are set, but these "
-                               "are not:\n\n"
-                               + "\n".join(f"  •  {m}" for m in missing)
-                               + "\n\nPreferences has a download link on each row — or press "
-                                 "Download models for me and point it at your models folder.",
-                     font=(FONT_FAMILY, 10), justify=tk.LEFT, wraplength=520,
-                     bg=COLORS["bg_deep"], fg=COLORS["text_explain"]).pack(
-                anchor=tk.W, pady=(8, 12))
-            row = tk.Frame(win, bg=COLORS["bg_deep"])
-            row.pack(anchor=tk.E)
-
-            def _to_prefs():
-                win.destroy()
-                self.notebook.select(self.prefs_tab)
-
-            def _never():
-                self.prefs["minimax_extras_prompt_dismissed"] = True
-                save_prefs(self.prefs)
-                win.destroy()
-
-            ttk.Button(row, text="Open Preferences", command=_to_prefs).pack(side=tk.LEFT)
-            ttk.Button(row, text="Later", command=win.destroy).pack(side=tk.LEFT, padx=(8, 0))
-            ttk.Button(row, text="Don't ask again", command=_never).pack(side=tk.LEFT,
-                                                                         padx=(8, 0))
-        self._check_minimax_extras = _check_minimax_extras     # bound for tests / re-checks
-        self.master.after(700, _check_minimax_extras)
-        # The audio-aware Training-tab rows (grey-outs, the voice-structure hint, the
-        # per-category retirement row) refresh on folder-change traces — but the RESTORED
-        # folder's trace fires during startup, before those widgets exist, and nothing
-        # re-fires after the tab is built. One deferred pass covers the restored state.
-        self.master.after(150, self._refresh_audio_only_ui)
-
-        # Tools card — highlights the post-training workbench tabs
-        tools_card = tk.Frame(container, bg=COLORS["bg_surface"],
-                              highlightbackground=COLORS["border"],
-                              highlightthickness=1, bd=0)
-        tools_card.pack(fill=tk.X, pady=(20, 0))
-
-        tk.Label(tools_card, text="Post-Training Tools",
-                 font=(FONT_FAMILY, 12, "bold"),
-                 fg=COLORS["text_primary"], bg=COLORS["bg_surface"]).pack(anchor=tk.W, padx=20, pady=(16, 2))
-        tk.Label(tools_card,
-                 text="Fizgig is more than a trainer — these tabs let you understand and tune any Klein LoRA "
-                      "you've made (or downloaded).",
-                 font=(FONT_FAMILY, 10),
-                 fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                 wraplength=760, justify=tk.LEFT).pack(anchor=tk.W, padx=20, pady=(0, 12))
-
-        tools = [
-            ("Profiler",
-             "Analyze a LoRA's per-block activation profile and produce an HTML report."),
-            ("Repair Studio",
-             "Live per-block sliders with side-by-side preview. Blend in a donor LoRA and "
-             "bake the result to a new .safetensors."),
-            ("LoRA the Explorer",
-             "Evolutionary discovery — the computer proposes random mutations, you pick favourites, "
-             "and the LoRA evolves. Seamlessly connected to Repair Studio."),
-            ("LoRA Royale",
-             "Compare epochs (or any LoRAs) on one seed, then export share-ready clips — seed, "
-             "prompt, and LoRA-strength travels, deflickered and ready for social."),
-            ("Extract",
-             "Distill a LoRA to a lower rank with optional block- and timestep-targeted presets. "
-             "Supports LyCORIS (LoKR / LoHa) sources."),
-        ]
-
-        for name, desc in tools:
-            row = tk.Frame(tools_card, bg=COLORS["bg_surface"])
-            row.pack(fill=tk.X, padx=20, pady=4)
-
-            tk.Label(row, text=name,
-                     font=(FONT_FAMILY, 11, "bold"),
-                     fg=COLORS["accent"], bg=COLORS["bg_surface"],
-                     width=16, anchor="w").pack(side=tk.LEFT, padx=(0, 12))
-            tk.Label(row, text=desc,
-                     font=(FONT_FAMILY, 10),
-                     fg=COLORS["text_secondary"], bg=COLORS["bg_surface"],
-                     anchor="w", justify=tk.LEFT, wraplength=620).pack(side=tk.LEFT, fill=tk.X, expand=True)
-
-        tk.Frame(tools_card, bg=COLORS["bg_surface"], height=12).pack()
-
-        self._add_youtube_help_button(scrollable_frame, "start", prominent=True)
-
-    def _browse_concept_folder(self):
-        """Folder picker for the second Multi Concept subject.
-
-        Deliberately NOT wired to image_folder_var: that one is the Start tab's, and Captions,
-        Image Prep, the Look filter, the gallery scorer and the loss watch all follow it."""
-        folder = filedialog.askdirectory(
-            initialdir=(self._concept_folder_vars[0].get()
-                        or self.image_folder_var.get()
-                        or self._pref_initialdir("input_dataset_dir")
-                        or os.getcwd()))
-        if folder:
-            self._concept_folder_vars[0].set(folder)
-
-    def _sync_distill_weight_state(self):
-        """Grey the teacher-weight box while identity-first is running the show.
-
-        In that mode phase 1 is teacher-ONLY (weight forced to 1.0) and phase 2 is
-        photographs-only, so the box changes nothing. Leaving it live invites people to tune a
-        dial that is not connected to anything."""
-        w = self.entries.get("MINIMAX_DISTILL_WEIGHT")
-        p1 = self.entries.get("MINIMAX_DISTILL_PHASE1")
-        if w is None or p1 is None:
-            return
-        _blended = str(p1.get()).startswith("Off")
-        try:
-            if _blended:
-                w.config(state="readonly")
-                # Hand back whatever the user had before identity-first took the dial over.
-                _stash = getattr(self, "_distill_weight_stash", None)
-                if _stash:
-                    w.set(_stash)
-                    self._distill_weight_stash = None
-            else:
-                # Show 1.0, because that is what phase 1 ACTUALLY runs at — leaving 0.8 sitting
-                # there greyed out told the user something untrue about their own run. (Phase 2
-                # then drops the teacher entirely; the log line says so at the switch.)
-                if str(w.get()) != "1.0" and not getattr(self, "_distill_weight_stash", None):
-                    self._distill_weight_stash = str(w.get())
-                w.set("1.0")
-                w.config(state="disabled")
-        except tk.TclError:
-            pass
-
-    # The recipe Multi Concept switches you into (Peter, 11 Aug). Applied ONCE, when the box is
-    # ticked — not locked. Locking caption dropout to a theory is what broke the last version of
-    # this mode, so these are starting points the user can still argue with.
-    # Deliberately NOT here: Adapter-relative LR. It is an LR strategy, and the two presets own
-    # that choice (Defaults on, Fast flat) - a box describing how your DATA is laid out has no
-    # business overruling the preset the user just loaded.
-    _MULTICONCEPT_DEFAULTS = {
-        "MINIMAX_CAPTION_DROPOUT": "0.10 (strong)",
-        "MINIMAX_DISTILL_REFS": "4",
-        "MINIMAX_DISTILL_PHASE1": "2 epochs",
-    }
-
-    def _warn_if_no_ref_dit(self):
-        """Identity-learn runs on ref2va. Say so when it is switched ON, not at Start.
-
-        validate_inputs already blocks the launch, but by then the user has captioned, cached
-        and pressed Start - and the remedy is a 21 GB download, so an hour of setup can be spent
-        before anything says the run cannot happen."""
-        if self._krea2_pref("minimax_ref_dit"):
-            return
-        messagebox.showinfo(
-            "Identity mode needs one more model",
-            "Learning identity from your dataset runs on the ref2va model - a separate 21 GB "
-            "file, and the only H3 build that accepts reference images. Fizgig does not "
-            "download it by default.\n\n"
-            "Preferences → MiniMax H3 → DiT (reference): paste the path if you "
-            "already have the file, or tick \"Include the reference DiT\" beside \"Download "
-            "models for me\" and let Fizgig fetch it.\n\n"
-            "Carry on setting the run up either way - Start will stop and remind you if the "
-            "path is still empty.")
-
-    def _on_minimax_distill_clicked(self):
-        """Only on a real click: setting the var programmatically must stay silent."""
-        if self.minimax_distill_var.get():
-            self._warn_if_no_ref_dit()
-
-    def _on_minimax_multiconcept_clicked(self):
-        """User CLICKED the box — apply the recipe, then refresh the rows.
-
-        Separate from _on_minimax_multiconcept_toggle because that one also runs on an
-        architecture switch and on every preset load; applying the recipe there would silently
-        overwrite settings the user had changed, every time they visited the tab."""
-        if self.minimax_multiconcept_var.get():
-            _changed = []
-            for _k, _v in self._MULTICONCEPT_DEFAULTS.items():
-                _w = self.entries.get(_k)
-                if _w is not None and str(_w.get()) != _v:
-                    _w.set(_v)
-                    _changed.append(f"{_k.replace('MINIMAX_', '').lower()}={_v}")
-            if hasattr(self, "minimax_distill_var") and not self.minimax_distill_var.get():
-                self.minimax_distill_var.set(True)
-                _changed.append("identity-learn=on")
-            if _changed:
-                self.update_console("[multi concept] applied: " + ", ".join(_changed)
-                                    + "  (all still editable)\n")
-        self._on_minimax_multiconcept_toggle()
-        self._sync_distill_weight_state()
-        if self.minimax_multiconcept_var.get() and self.minimax_distill_var.get():
-            self._warn_if_no_ref_dit()
-
-    def _on_minimax_multiconcept_toggle(self):
-        """Show the extra folder row. Caption dropout is deliberately NOT touched.
-
-        It used to be forced off here, on the theory that training a few percent of steps against
-        the EMPTY prompt teaches the model to produce a subject with no trigger — the mechanism by
-        which two subjects bleed. Peter's own A/B said otherwise (11 Aug): with distillation off,
-        one folder WITH dropout beat two folders without it. Whatever dropout costs in bleed, it
-        appears to be worth more as regularisation at this scale. The dial goes back to the user
-        rather than being locked to a theory the data does not support."""
-        on = bool(self.minimax_multiconcept_var.get()) and self._is_minimax_arch()
-        for w in (getattr(self, "_minimax_mc_dir_frame", None),
-                  getattr(self, "_minimax_mc_hint", None)):
-            if w is not None:
-                self._set_widget_visible(w, on)
-        # The "no reference steering" warning only makes sense in multi-concept with distill off.
-        _nd = getattr(self, "_minimax_mc_nodistill_hint", None)
-        if _nd is not None:
-            _distill = bool(getattr(self, "minimax_distill_var", None)
-                            and self.minimax_distill_var.get())
-            self._set_widget_visible(_nd, on and not _distill)
-
-    def _browse_image_folder(self):
-        """Folder picker for the Start tab (unified image folder).
-
-        Falls back through: the folder you last used, then the Preferences default (which the pod
-        image seeds to /workspace/datasets so Browse opens where uploads land), then cwd."""
-        folder = filedialog.askdirectory(
-            initialdir=(self.image_folder_var.get()
-                        or self._pref_initialdir("input_dataset_dir")
-                        or os.getcwd()))
-        if folder:
-            self.image_folder_var.set(folder)
 
     def create_training_settings(self):
         """Create the Training tab (Start-tab styled)."""
@@ -14355,2222 +13944,6 @@ class LoRATrainerGUI(StylingMixin, TabScaffoldMixin, ConsoleValidationMixin):
                 pass
             self._explorer_engine = None
             self.explorer_status_var.set("Models unloaded (tab switch). Load a LoRA to resume.")
-
-    # endregion
-
-    # region Extract Tab
-
-    def create_extract_tab(self):
-        """Create the Extract tab (Start-tab styled) — activation-based LoRA extraction."""
-        scrollable_frame, _ = self.create_scrollable_frame(self.extract_tab)
-
-        outer = tk.Frame(scrollable_frame, bg=COLORS["bg_deep"])
-        outer.pack(fill=tk.BOTH, expand=True)
-
-        self._add_tab_banner(
-            outer,
-            "Extract",
-            "Distill an existing LoRA down to a lower rank. Klein: block + timestep targeting, optional "
-            "activation-weighted SVD. Krea 2 and MiniMax H3: pure weight SVD over all blocks (no block map yet).",
-        )
-
-        # Model family selector. Krea 2 / MiniMax H3 = pure weight SVD over all blocks (no pipeline /
-        # prompt / timesteps / block presets), so those cards are hidden for both.
-        _efam = str(self.last_used.get("extract_family", "klein"))
-        if _efam not in ("klein", "krea2", "minimax"):
-            _efam = "klein"
-        self.extract_family_var = tk.StringVar(value=_efam)
-        efam_card = self._start_section_card(
-            outer, "Model Family",
-            "Klein 9B (full extractor), Krea 2 or MiniMax H3 (weight-only SVD; block-targeting presets "
-            "come once each block map exists). Browsing a LoRA auto-switches to its family.",
-        )
-        _ef = tk.Frame(efam_card, bg=COLORS["bg_surface"])
-        _ef.pack(anchor=tk.W)
-        ttk.Radiobutton(_ef, text="Klein 9B", variable=self.extract_family_var, value="klein",
-                        command=self._on_extract_family_changed).pack(side=tk.LEFT, padx=(0, 20))
-        ttk.Radiobutton(_ef, text="Krea 2", variable=self.extract_family_var, value="krea2",
-                        command=self._on_extract_family_changed).pack(side=tk.LEFT, padx=(0, 20))
-        ttk.Radiobutton(_ef, text="MiniMax H3", variable=self.extract_family_var, value="minimax",
-                        command=self._on_extract_family_changed).pack(side=tk.LEFT)
-
-        # Card 1: Source & Output
-        io_card = self._start_section_card(
-            outer, "Source & Output",
-            "Choose the source LoRA and name the extraction — it will land in your LoRA output folder.",
-        )
-        io_card.grid_columnconfigure(1, weight=1)
-
-        ttk.Label(io_card, text="Source LoRA:").grid(row=0, column=0, sticky=tk.W, padx=(0, 10), pady=4)
-        self.extract_source_var = tk.StringVar()
-        ttk.Entry(io_card, textvariable=self.extract_source_var, width=50).grid(row=0, column=1, sticky=tk.EW, pady=4)
-        ttk.Button(io_card, text="Browse", command=self._browse_extract_source).grid(row=0, column=2, sticky=tk.W, padx=(8, 0), pady=4)
-
-        ttk.Label(io_card, text="Output Name:").grid(row=1, column=0, sticky=tk.W, padx=(0, 10), pady=4)
-        self.extract_output_var = tk.StringVar()
-        ttk.Entry(io_card, textvariable=self.extract_output_var, width=50).grid(row=1, column=1, sticky=tk.EW, pady=4)
-        tk.Label(io_card, text="(will be saved in your LoRA output folder)",
-                 font=(FONT_FAMILY, 9, "italic"),
-                 fg=COLORS["text_muted"], bg=COLORS["bg_surface"]).grid(
-            row=2, column=1, sticky=tk.W, pady=(0, 4)
-        )
-
-        # Card 2: Preset
-        preset_card = self._start_section_card(
-            outer, "Preset",
-            "Fast * = pure weight SVD (samples=0, no GPU probes).  |  Identity = single 1-16.  |  "
-            "Style = style+comp @ late timesteps.  |  Style+Composition = double 0-7 + single 0-1 + single 2 @ 0.5.  |  "
-            "Details = single 12-23.",
-        )
-        self._extract_preset_container = preset_card.master.master
-        preset_row = tk.Frame(preset_card, bg=COLORS["bg_surface"])
-        preset_row.pack(anchor=tk.W)
-        ttk.Label(preset_row, text="Extract Preset:").pack(side=tk.LEFT, padx=(0, 10))
-        self.extract_preset_var = tk.StringVar(value="Identity")
-        preset_combo = ttk.Combobox(
-            preset_row, textvariable=self.extract_preset_var,
-            values=["All Blocks", "Fast SVD", "Identity", "Fast Identity", "Style",
-                    "Style+Composition", "Fast Style+Composition", "Details", "Fast Details", "Custom"],
-            state="readonly", width=24,
-        )
-        preset_combo.pack(side=tk.LEFT)
-        preset_combo.bind("<<ComboboxSelected>>", self._on_extract_preset_changed)
-
-        # Card 3: Custom Blocks — pack-managed so pack_forget/pack drives visibility
-        self._extract_custom_frame = tk.Frame(outer, bg=COLORS["bg_deep"])
-        self._extract_custom_frame.pack(fill=tk.X, padx=36, pady=(0, 16))
-        custom_card = tk.Frame(self._extract_custom_frame, bg=COLORS["bg_surface"],
-                               highlightbackground=COLORS["border"], highlightthickness=1, bd=0)
-        custom_card.pack(fill=tk.X)
-        tk.Label(custom_card, text="Custom Blocks",
-                 font=(FONT_FAMILY, 12, "bold"),
-                 fg=COLORS["text_primary"], bg=COLORS["bg_surface"]).pack(anchor=tk.W, padx=20, pady=(16, 2))
-        tk.Label(custom_card,
-                 text="Pick individual blocks to target. Only shown when preset = Custom.",
-                 font=(FONT_FAMILY, 10),
-                 fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                 wraplength=760, justify=tk.LEFT).pack(anchor=tk.W, padx=20, pady=(0, 10))
-
-        custom_content = tk.Frame(custom_card, bg=COLORS["bg_surface"])
-        custom_content.pack(fill=tk.X, padx=20, pady=(0, 16))
-
-        self.extract_block_vars = {}
-
-        header_row = tk.Frame(custom_content, bg=COLORS["bg_surface"])
-        header_row.pack(anchor=tk.W, pady=(0, 6))
-        tk.Label(header_row, text="Select individual blocks:",
-                 font=(FONT_FAMILY, 10, "bold"),
-                 fg=COLORS["text_primary"], bg=COLORS["bg_surface"]).pack(side=tk.LEFT, padx=(0, 12))
-        ttk.Button(header_row, text="All", width=5,
-                   command=lambda: self._set_all_extract_blocks(True)).pack(side=tk.LEFT, padx=2)
-        ttk.Button(header_row, text="None", width=5,
-                   command=lambda: self._set_all_extract_blocks(False)).pack(side=tk.LEFT, padx=2)
-        ttk.Button(header_row, text="Identity", width=8,
-                   command=lambda: self._set_category_extract_blocks("identity")).pack(side=tk.LEFT, padx=2)
-        ttk.Button(header_row, text="Style+Comp", width=10,
-                   command=lambda: self._set_category_extract_blocks("style_composition")).pack(side=tk.LEFT, padx=2)
-        ttk.Button(header_row, text="Details", width=8,
-                   command=lambda: self._set_category_extract_blocks("details")).pack(side=tk.LEFT, padx=2)
-
-        double_row = tk.Frame(custom_content, bg=COLORS["bg_surface"])
-        double_row.pack(anchor=tk.W, pady=2)
-        tk.Label(double_row, text="Double:", width=8, fg="#5B9BD5", bg=COLORS["bg_surface"],
-                 font=(FONT_FAMILY, 10)).pack(side=tk.LEFT)
-        for i in range(8):
-            key = f"double_blocks.{i}"
-            var = tk.BooleanVar(value=False)
-            self.extract_block_vars[key] = var
-            ttk.Checkbutton(double_row, text=str(i), variable=var).pack(side=tk.LEFT, padx=2)
-
-        single_row1 = tk.Frame(custom_content, bg=COLORS["bg_surface"])
-        single_row1.pack(anchor=tk.W, pady=2)
-        tk.Label(single_row1, text="Single:", width=8, fg="#70AD47", bg=COLORS["bg_surface"],
-                 font=(FONT_FAMILY, 10)).pack(side=tk.LEFT)
-        for i in range(12):
-            key = f"single_blocks.{i}"
-            var = tk.BooleanVar(value=False)
-            self.extract_block_vars[key] = var
-            ttk.Checkbutton(single_row1, text=str(i), variable=var).pack(side=tk.LEFT, padx=2)
-
-        single_row2 = tk.Frame(custom_content, bg=COLORS["bg_surface"])
-        single_row2.pack(anchor=tk.W, pady=2)
-        tk.Label(single_row2, text="Single:", width=8, fg="#ED7D31", bg=COLORS["bg_surface"],
-                 font=(FONT_FAMILY, 10)).pack(side=tk.LEFT)
-        for i in range(12, 24):
-            key = f"single_blocks.{i}"
-            var = tk.BooleanVar(value=False)
-            self.extract_block_vars[key] = var
-            ttk.Checkbutton(single_row2, text=str(i), variable=var).pack(side=tk.LEFT, padx=2)
-
-        tk.Label(custom_content,
-                 text="double + single 0-1 = style+composition  |  single 1-16 = identity (overlaps at 1 and 12-16)  |  single 12-23 = details",
-                 font=(FONT_FAMILY, 9, "italic"),
-                 fg=COLORS["text_muted"], bg=COLORS["bg_surface"]).pack(anchor=tk.W, pady=(4, 0))
-
-        self._extract_custom_frame.pack_forget()  # hidden until preset = Custom
-
-        # Card 4: Options (rank / timesteps / forward passes)
-        options_card = self._start_section_card(
-            outer, "Options",
-            "Timesteps: 'all' for general, 'late' for style, 'early' for composition.  |  "
-            "Forward Passes: 0 = pure weight SVD (fastest, timestep-agnostic). Higher = better activation-weighted accuracy.",
-        )
-        # Anchor for the Custom Blocks card's re-pack so it reappears between Preset and Options.
-        self._extract_options_anchor = options_card.master.master
-        options_row = tk.Frame(options_card, bg=COLORS["bg_surface"])
-        options_row.pack(anchor=tk.W)
-
-        ttk.Label(options_row, text="Target Rank:").pack(side=tk.LEFT, padx=(0, 6))
-        self.extract_rank_var = tk.StringVar(value="4")
-        rank_combo = ttk.Combobox(options_row, textvariable=self.extract_rank_var,
-                     values=["1", "2", "4", "8", "16"], state="readonly", width=4)
-        rank_combo.pack(side=tk.LEFT, padx=(0, 20))
-        rank_combo.bind("<<ComboboxSelected>>", lambda e: self._update_extract_output_name())
-
-        self._extract_timesteps_label = ttk.Label(options_row, text="Timesteps:")
-        self._extract_timesteps_label.pack(side=tk.LEFT, padx=(0, 6))
-        self.extract_timesteps_var = tk.StringVar(value="all")
-        self._extract_timesteps_combo = ttk.Combobox(
-            options_row, textvariable=self.extract_timesteps_var,
-            values=["all", "early", "mid", "late"], state="readonly", width=8,
-        )
-        self._extract_timesteps_combo.pack(side=tk.LEFT, padx=(0, 20))
-
-        self._extract_samples_label = ttk.Label(options_row, text="Forward Passes:")
-        self._extract_samples_label.pack(side=tk.LEFT, padx=(0, 6))
-        self.extract_samples_var = tk.StringVar(value="16")
-        self._extract_samples_combo = ttk.Combobox(
-            options_row, textvariable=self.extract_samples_var,
-            values=["0", "8", "16", "32"], state="readonly", width=4,
-        )
-        self._extract_samples_combo.pack(side=tk.LEFT)
-        self._extract_samples_combo.bind("<<ComboboxSelected>>", self._on_extract_samples_changed)
-
-        # Card 5: Prompt
-        prompt_card = self._start_section_card(
-            outer, "Prompt",
-            "Used during the GPU probe forward passes. Include the source LoRA's trigger word for best results.",
-        )
-        self._extract_prompt_container = prompt_card.master.master
-        prompt_card.grid_columnconfigure(1, weight=1)
-        ttk.Label(prompt_card, text="Prompt:").grid(row=0, column=0, sticky=tk.W, padx=(0, 10), pady=4)
-        self.extract_prompt_var = tk.StringVar(value="")
-        ttk.Entry(prompt_card, textvariable=self.extract_prompt_var, width=50).grid(row=0, column=1, sticky=tk.EW, pady=4)
-
-        # Card 6: Run
-        run_card = self._start_section_card(
-            outer, "Run",
-            "Extraction runs SVD on each block and can take several minutes depending on rank and block count.",
-        )
-        self._extract_run_container = run_card.master.master
-        run_row = tk.Frame(run_card, bg=COLORS["bg_surface"])
-        run_row.pack(anchor=tk.W)
-        self.extract_run_btn = ttk.Button(run_row, text="Extract LoRA", command=self._run_extract, style="Primary.TButton")
-        self.extract_run_btn.pack(side=tk.LEFT, padx=(0, 12))
-        self.extract_open_btn = ttk.Button(run_row, text="Open Output Folder", command=self._open_extract_folder, state="disabled")
-        self.extract_open_btn.pack(side=tk.LEFT)
-
-        self.extract_progress_var = tk.StringVar(value="")
-        tk.Label(run_card, textvariable=self.extract_progress_var,
-                 font=(FONT_FAMILY, 10, "bold"),
-                 fg=COLORS["accent_hover"], bg=COLORS["bg_surface"]).pack(anchor=tk.W, pady=(10, 0))
-
-        # Family-aware "how long this takes" note (set by _apply_extract_family_ui).
-        self.extract_time_note_var = tk.StringVar(value="")
-        tk.Label(run_card, textvariable=self.extract_time_note_var,
-                 font=(FONT_FAMILY, 10), fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                 wraplength=760, justify=tk.LEFT).pack(anchor=tk.W, pady=(8, 0))
-
-        # Card 7: Output Log
-        log_card = self._start_section_card(outer, "Output Log", None)
-        self.extract_log = scrolledtext.ScrolledText(
-            log_card, height=14, width=80,
-            bg=COLORS["bg_deep"], fg=COLORS["text_primary"],
-            wrap="word", state="disabled",
-            relief="flat", highlightthickness=1,
-            highlightbackground=COLORS["border"], highlightcolor=COLORS["border_focus"],
-        )
-        self.extract_log.pack(fill=tk.BOTH, expand=True)
-
-        self._extract_output_path = None
-
-        # Apply the persisted family (krea2 hides the Klein-only block/prompt/probe controls).
-        self._apply_extract_family_ui(str(self.extract_family_var.get()) == "krea2")
-
-        self._add_youtube_help_button(outer, "extract")
-
-    def _on_mining_preset_changed(self, *args):
-        """Fill gradient mining controls from the selected preset."""
-        name = self._mining_preset_var.get()
-        preset = self._mining_presets.get(name)
-        if not preset:
-            return
-        for key, entry_key in [("snr", "GRADIENT_MINING_THRESHOLD"),
-                                ("ortho", "GRADIENT_MINING_ORTHO")]:
-            entry = self.entries.get(entry_key)
-            if entry and key in preset:
-                entry.delete(0, tk.END)
-                entry.insert(0, preset[key])
-
-    def _on_training_preset_changed(self, *args):
-        """Auto-fill MIN/MAX_TIMESTEP and show/hide custom block picker based on training preset."""
-        preset = self.training_preset_var.get()
-
-        # Show/hide custom block panel
-        if preset == "Custom":
-            self._training_custom_frame.grid()
-        else:
-            self._training_custom_frame.grid_remove()
-
-        # Skip timestep auto-fill for Custom (user-driven)
-        if preset == "Custom":
-            self._last_area_applied = preset
-            return
-
-        # Only rewrite MIN/MAX when the Model Area actually CHANGED. This handler is also
-        # invoked as a visibility refresh (arch switches, tab builds) — unconditionally
-        # auto-filling wiped the user's hand-set timesteps on every one of those calls.
-        if getattr(self, "_last_area_applied", None) == preset:
-            return
-        self._last_area_applied = preset
-
-        # Auto-fill MIN/MAX_TIMESTEP entries
-        min_entry = self.entries.get("MIN_TIMESTEP")
-        max_entry = self.entries.get("MAX_TIMESTEP")
-        if min_entry is None or max_entry is None:
-            return
-
-        if preset == "Style":
-            min_val, max_val = "0", "400"
-        else:  # Full Model, Identity, Style+Composition, Details
-            min_val, max_val = "", ""
-
-        min_entry.delete(0, tk.END)
-        if min_val:
-            min_entry.insert(0, min_val)
-        max_entry.delete(0, tk.END)
-        if max_val:
-            max_entry.insert(0, max_val)
-        self.settings["MIN_TIMESTEP"] = min_val
-        self.settings["MAX_TIMESTEP"] = max_val
-        if hasattr(self, "_update_noise_range_label"):
-            try:
-                self._update_noise_range_label()
-            except Exception:
-                pass
-
-    def _set_all_training_blocks(self, value: bool):
-        """Tick/untick all per-block checkboxes in Custom training mode."""
-        for var in self.training_block_vars.values():
-            var.set(value)
-
-    def _set_category_training_blocks(self, category: str):
-        """Select all blocks in a category (identity/style_composition/details). Clears others."""
-        for key, var in self.training_block_vars.items():
-            var.set(False)
-        if category == "style_composition":
-            for i in range(8):
-                self.training_block_vars[f"double_blocks.{i}"].set(True)
-            for i in (0, 1):
-                self.training_block_vars[f"single_blocks.{i}"].set(True)
-        elif category == "identity":
-            for i in range(1, 17):
-                self.training_block_vars[f"single_blocks.{i}"].set(True)
-        elif category == "details":
-            for i in range(12, 24):
-                self.training_block_vars[f"single_blocks.{i}"].set(True)
-
-    def _build_custom_training_patterns(self):
-        """Build a list of include_patterns regexes from the Custom block checkboxes."""
-        selected = [key for key, var in self.training_block_vars.items() if var.get()]
-        if not selected:
-            return None
-        patterns = []
-        for key in selected:
-            # key is "double_blocks.N" or "single_blocks.N"
-            kind, idx = key.split(".")
-            patterns.append(rf".*{kind}\.{idx}\..*")
-        return patterns
-
-    def _on_extract_preset_changed(self, *args):
-        """Show/hide custom block checkboxes and auto-switch timesteps/samples based on preset."""
-        preset = self.extract_preset_var.get()
-        if preset == "Custom":
-            # Pack before Options so the custom card appears between Preset and Options.
-            self._extract_custom_frame.pack(fill=tk.X, padx=36, pady=(0, 16),
-                                             before=self._extract_options_anchor)
-        else:
-            self._extract_custom_frame.pack_forget()
-
-        # Fast * presets force samples=0 (pure weight SVD)
-        is_fast = preset in ("Fast SVD", "Fast Identity", "Fast Style+Composition", "Fast Details")
-        if is_fast:
-            self.extract_samples_var.set("0")
-        elif preset in ("All Blocks", "Identity", "Style", "Style+Composition", "Details"):
-            # Activation-weighted presets need samples>0; bump off 0 if user is coming from a Fast variant
-            if self.extract_samples_var.get() == "0":
-                self.extract_samples_var.set("16")
-
-        # Timestep auto-fill
-        if preset == "Style":
-            self.extract_timesteps_var.set("late")
-        elif preset in ("All Blocks", "Fast SVD", "Identity", "Fast Identity",
-                        "Style+Composition", "Fast Style+Composition", "Details", "Fast Details"):
-            self.extract_timesteps_var.set("all")
-
-        # Reflect timestep combo state (samples may have just changed above)
-        self._apply_extract_samples_state()
-
-        # Update suggested output filename to match new preset
-        self._update_extract_output_name()
-
-    def _on_extract_samples_changed(self, *args):
-        """Grey out timesteps when samples=0; map presets to their Fast variant where applicable."""
-        if self.extract_samples_var.get() == "0":
-            preset = self.extract_preset_var.get()
-            # Map activation-weighted presets to their Fast equivalent so the UI reflects reality
-            fast_map = {
-                "All Blocks": "Fast SVD",
-                "Identity": "Fast Identity",
-                "Style": "Fast Style+Composition",   # Style is inherently timestep-based; Fast loses that meaning
-                "Style+Composition": "Fast Style+Composition",
-                "Details": "Fast Details",
-            }
-            if preset in fast_map:
-                self.extract_preset_var.set(fast_map[preset])
-                self.extract_timesteps_var.set("all")
-        self._apply_extract_samples_state()
-
-    def _apply_extract_samples_state(self):
-        """Timesteps dropdown is meaningful only when forward passes > 0."""
-        if self.extract_samples_var.get() == "0":
-            self._extract_timesteps_combo.configure(state="disabled")
-        else:
-            self._extract_timesteps_combo.configure(state="readonly")
-
-    def _set_all_extract_blocks(self, value: bool):
-        """Tick/untick all per-block checkboxes in Custom mode."""
-        for var in self.extract_block_vars.values():
-            var.set(value)
-
-    def _set_category_extract_blocks(self, category: str):
-        """Select all blocks in a category (identity/style_composition/details). Clears others."""
-        for key, var in self.extract_block_vars.items():
-            var.set(False)
-        if category == "style_composition":
-            # double 0-7 + single 0-1 + single 2 (at full strength in Custom mode)
-            for i in range(8):
-                self.extract_block_vars[f"double_blocks.{i}"].set(True)
-            for i in (0, 1, 2):
-                self.extract_block_vars[f"single_blocks.{i}"].set(True)
-        elif category == "identity":
-            for i in range(1, 17):
-                self.extract_block_vars[f"single_blocks.{i}"].set(True)
-        elif category == "details":
-            for i in range(12, 24):
-                self.extract_block_vars[f"single_blocks.{i}"].set(True)
-
-    def _update_extract_output_name(self):
-        """Regenerate the suggested output filename from source + preset + rank."""
-        source = self.extract_source_var.get().strip()
-        if not source:
-            return
-        base = os.path.splitext(os.path.basename(source))[0]
-        preset_slug = self.extract_preset_var.get().lower().replace("+", "_").replace(" ", "_")
-        self.extract_output_var.set(f"{base}_{preset_slug}_r{self.extract_rank_var.get()}.safetensors")
-
-    def _browse_extract_source(self):
-        filepath = filedialog.askopenfilename(
-            title="Select source LoRA",
-            filetypes=[("SafeTensors", "*.safetensors"), ("All files", "*.*")],
-            initialdir=self._lora_initialdir(),
-        )
-        if filepath:
-            self.extract_source_var.set(filepath)
-            self._update_extract_output_name()
-            # Header-only family sniff — picking a LoRA from the wrong family auto-switches
-            # instead of erroring at run time (same as Explorer / Profiler).
-            try:
-                from fizgig.networks.lora import lora_family_from_file
-                fam = lora_family_from_file(filepath)
-                if fam in ("klein", "krea2", "minimax") and fam != self.extract_family_var.get():
-                    self.extract_family_var.set(fam)
-                    self._on_extract_family_changed()
-            except Exception:
-                pass
-
-    def _extract_log(self, text):
-        """Append to extract log (preserves user scroll position). Marshals to the main
-        thread — the extract worker calls this from its own thread."""
-        import threading
-        if threading.current_thread() is not threading.main_thread():
-            self.master.after(0, self._extract_log, text)
-            return
-        self._append_global_log(text)
-        self._smart_text_insert(self.extract_log, text)
-
-    def _open_extract_folder(self):
-        if self._extract_output_path and os.path.exists(self._extract_output_path):
-            folder = os.path.dirname(self._extract_output_path)
-            self._open_in_file_manager(folder)
-
-    def _run_extract(self):
-        """Start extraction in a background thread."""
-        if str(self.extract_family_var.get()) in ("krea2", "minimax"):
-            self._run_extract_krea2()      # weight-only path; model-agnostic, serves H3 too
-            return
-
-        source = self.extract_source_var.get()
-        if not source or not os.path.exists(source):
-            messagebox.showerror("Error", "Please select a valid source LoRA.")
-            return
-
-        output_name = self.extract_output_var.get().strip()
-        if not output_name:
-            messagebox.showerror("Error", "Please enter an output name.")
-            return
-        if not output_name.endswith(".safetensors"):
-            output_name += ".safetensors"
-
-        # Fast (weight-only, samples=0) presets run entirely from the safetensors file:
-        # no pipeline is loaded, so don't demand a DiT/VAE/TE or a prompt they never use
-        # (the CLI already worked without them).
-        try:
-            _samples = int(self.extract_samples_var.get().strip() or 0)
-        except (ValueError, AttributeError):
-            _samples = 0
-
-        prompt = self.extract_prompt_var.get().strip()
-        dit_path = self.prefs_vars["distilled_dit"].get()
-        vae_path = self.prefs_vars["vae"].get()
-        te_path = self.prefs_vars["text_encoder"].get()
-
-        if _samples > 0:
-            if not prompt:
-                messagebox.showerror("Error", "Please enter a prompt (trigger word recommended).")
-                return
-            for path, name in [(dit_path, "DiT"), (vae_path, "VAE"), (te_path, "Text Encoder")]:
-                if not path or not os.path.exists(path):
-                    messagebox.showerror("Error", f"{name} not found:\n{path}\n\nCheck Preferences tab.")
-                    return
-
-        # Build block list from preset (or custom individual blocks)
-        preset = self.extract_preset_var.get()
-        custom_blocks = None  # Per-block list for Custom mode
-        if preset in ("Identity", "Fast Identity"):
-            blocks = ["identity"]
-        elif preset in ("Style", "Style+Composition", "Fast Style+Composition"):
-            blocks = ["style_composition"]
-        elif preset in ("Details", "Fast Details"):
-            blocks = ["details"]
-        elif preset in ("All Blocks", "Fast SVD"):
-            blocks = ["all"]
-        else:  # Custom
-            selected = [key for key, var in self.extract_block_vars.items() if var.get()]
-            if not selected:
-                messagebox.showerror("Error", "Custom mode: please select at least one block.")
-                return
-            blocks = ["custom"]
-            custom_blocks = selected
-
-        output_dir = self.prefs_vars["lora_output_dir"].get()
-        os.makedirs(output_dir, exist_ok=True)
-        output_path = os.path.join(output_dir, output_name)
-        # Never silently overwrite: the name is built from source+preset+rank only, so two
-        # different Custom selections at the same rank land on the same filename.
-        if os.path.exists(output_path):
-            _stem, _ext = os.path.splitext(output_path)
-            _n = 2
-            while os.path.exists(f"{_stem}_{_n}{_ext}"):
-                _n += 1
-            output_path = f"{_stem}_{_n}{_ext}"
-
-        # Disable button, clear log
-        self.extract_run_btn.configure(state="disabled")
-        self.extract_open_btn.configure(state="disabled")
-        self.extract_log.configure(state="normal")
-        self.extract_log.delete(1.0, tk.END)
-        self.extract_log.configure(state="disabled")
-        self.extract_progress_var.set("Loading models...")
-
-        import threading
-        thread = threading.Thread(
-            target=self._extract_worker,
-            args=(source, output_path, dit_path, vae_path, te_path, blocks, prompt, custom_blocks),
-            daemon=True,
-        )
-        thread.start()
-
-    def _extract_worker(self, source, output_path, dit_path, vae_path, te_path, blocks, prompt, custom_blocks=None):
-        """Background worker for extraction."""
-        try:
-            import sys
-            sys.path.insert(0, os.path.join(os.path.dirname(__file__), "src"))
-
-            from fizgig.extraction.extractor import LoRAExtractor, ExtractionConfig
-
-            timestep_presets = {
-                "all": (0.0, 1.0),
-                "early": (0.6, 1.0),
-                "mid": (0.3, 0.7),
-                "late": (0.0, 0.4),
-            }
-
-            rank = int(self.extract_rank_var.get())
-            samples = int(self.extract_samples_var.get())
-            timesteps = timestep_presets[self.extract_timesteps_var.get()]
-
-            config = ExtractionConfig(
-                source_lora_path=source,
-                output_lora_path=output_path,
-                target_rank=rank,
-                timestep_range=timesteps,
-                include_blocks=blocks,
-                custom_blocks=custom_blocks,
-                num_samples=samples,
-                prompt=prompt,
-                width=1024,
-                height=1024,
-                seed=42,
-            )
-
-            def progress(stage, current, total):
-                def _update():
-                    self.extract_progress_var.set(f"{stage}: {current+1}/{total}")
-                self.master.after(0, _update)
-
-            pipeline = None
-            if samples == 0:
-                # Pure weight SVD — no pipeline needed, no GPU models loaded
-                self.master.after(0, lambda: self._extract_log(f"Pure weight SVD: blocks={blocks}, rank={rank}\n"))
-                result = LoRAExtractor.extract_weight_only(config, progress_callback=progress)
-            else:
-                # Activation-weighted SVD — needs full pipeline for forward passes
-                from fizgig.klein.inference import KleinInferencePipeline
-
-                # Auto-detect model version and fp8 from filename
-                dit_basename = os.path.basename(dit_path).lower()
-                model_version = "klein-base-9b" if "base" in dit_basename else "klein-9b"
-                is_fp8_model = "fp8" in dit_basename
-
-                self.master.after(0, lambda: self._extract_log("Loading models...\n"))
-
-                pipeline = KleinInferencePipeline()
-                pipeline.load_models(
-                    dit_path=dit_path,
-                    vae_path=vae_path,
-                    text_encoder_path=te_path,
-                    model_version=model_version,
-                    device="cuda",
-                    fp8_scaled=not is_fp8_model,
-                    fp8_text_encoder=True,
-                    blocks_to_swap=self._get_inference_blocks_to_swap(),
-                )
-
-                self.master.after(0, lambda: self._extract_log(
-                    f"Starting extraction: blocks={blocks}, rank={rank}, samples={samples}\n"))
-
-                extractor = LoRAExtractor(pipeline)
-                result = extractor.extract(config, progress_callback=progress)
-
-            if pipeline is not None:
-                pipeline.unload_models()
-
-            # A valid-but-empty artifact is a failure, not a success: a Details preset on a
-            # Style LoRA used to produce a ~340-byte tensorless file and "Extraction
-            # complete! Layers extracted: 0".
-            if result.num_layers_extracted == 0:
-                try:
-                    os.remove(result.output_path)
-                except OSError:
-                    pass
-                raise RuntimeError(
-                    f"Extraction produced 0 layers — the selected blocks ({blocks}) don't "
-                    f"exist in this LoRA. Run it through the Profiler to see which blocks "
-                    f"it actually trains, then pick a matching preset.")
-
-            summary = f"\nExtraction complete!\n"
-            summary += f"  Output: {result.output_path}\n"
-            summary += f"  Layers extracted: {result.num_layers_extracted}\n"
-            summary += f"  Target rank: {result.target_rank}\n"
-            summary += f"  Total params: {result.total_params:,}\n"
-            summary += f"  Time: {result.elapsed_seconds:.1f}s\n"
-
-            self._extract_output_path = output_path
-
-            def _update_ui():
-                self._extract_log(summary)
-                self.extract_progress_var.set("Done!")
-                self.extract_run_btn.configure(state="normal")
-                self.extract_open_btn.configure(state="normal")
-
-            self.master.after(0, _update_ui)
-
-        except Exception as e:
-            import traceback
-            error_msg = f"Extraction failed:\n{traceback.format_exc()}"
-            def _show_error():
-                self._extract_log(error_msg)
-                self.extract_progress_var.set("Error")
-                self.extract_run_btn.configure(state="normal")
-            self.master.after(0, _show_error)
-
-    # --- Krea 2 extract (weight-only SVD over all blocks; no pipeline / prompt / block map) ---
-
-    def _on_extract_family_changed(self):
-        fam = str(self.extract_family_var.get())
-        if fam not in ("klein", "krea2", "minimax"):
-            fam = "klein"
-        self.last_used["extract_family"] = fam
-        self._save_last_used_paths()
-        self._apply_extract_family_ui(fam != "klein")
-
-    def _apply_extract_family_ui(self, is_krea2):
-        """Krea 2 / MiniMax H3 mode: pure weight SVD over all blocks. Hide the block-preset,
-        custom-block, prompt and activation-probe (timesteps + forward passes) controls — only
-        Target Rank plus Source/Output/Run remain. Klein mode restores everything.
-        (`is_krea2` is historical naming: True means any weight-only family.)"""
-        # (widget, original padx) — restored verbatim so the klein branch is idempotent.
-        probe_widgets = [
-            (getattr(self, "_extract_timesteps_label", None), (0, 6)),
-            (getattr(self, "_extract_timesteps_combo", None), (0, 20)),
-            (getattr(self, "_extract_samples_label", None), (0, 6)),
-            (getattr(self, "_extract_samples_combo", None), (0, 0)),
-        ]
-        if is_krea2:
-            for c in (getattr(self, "_extract_preset_container", None),
-                      getattr(self, "_extract_prompt_container", None)):
-                if c is not None:
-                    c.pack_forget()
-            if getattr(self, "_extract_custom_frame", None) is not None:
-                self._extract_custom_frame.pack_forget()
-            for w, _ in probe_widgets:
-                if w is not None:
-                    w.pack_forget()
-            # Force weight-only all-blocks regardless of stale Klein selections.
-            self.extract_samples_var.set("0")
-            self.extract_timesteps_var.set("all")
-            if str(self.extract_family_var.get()) == "minimax":
-                self.extract_time_note_var.set(
-                    "MiniMax H3 is a 33B model - weight SVD runs over every trained module "
-                    "(208+ Linears, up to 5376 wide). Expect several minutes on a free GPU. "
-                    "If the GPU is busy (a training run, ComfyUI, another preview), each SVD "
-                    "falls back to the CPU and runs much slower - free up VRAM first.")
-                return
-            self.extract_time_note_var.set(
-                "⏱ Krea 2 is a 12.9B model — weight SVD runs over all 264 modules, several of "
-                "them very large (e.g. 36864×6144). Expect roughly 5–10 minutes on a free GPU. "
-                "If the GPU is busy (a training run, ComfyUI, another preview), each SVD falls back "
-                "to the CPU and the whole run can take 60 min+ — free up VRAM first for the fast path.")
-        else:
-            anchor = getattr(self, "_extract_options_anchor", None)
-            if getattr(self, "_extract_preset_container", None) is not None and anchor is not None:
-                self._extract_preset_container.pack(fill=tk.X, padx=36, pady=(0, 16), before=anchor)
-            run_anchor = getattr(self, "_extract_run_container", None)
-            if getattr(self, "_extract_prompt_container", None) is not None and run_anchor is not None:
-                self._extract_prompt_container.pack(fill=tk.X, padx=36, pady=(0, 16), before=run_anchor)
-            # Re-pack probe widgets in their original order/padding after the rank combo.
-            for w, padx in probe_widgets:
-                if w is not None:
-                    w.pack(side=tk.LEFT, padx=padx)
-            # Custom-block frame visibility is owned by the preset combo; refresh it.
-            if hasattr(self, "_on_extract_preset_changed"):
-                self._on_extract_preset_changed()
-            self.extract_time_note_var.set(
-                "⏱ Fast SVD presets (weight-only) finish in well under a minute. Activation-weighted "
-                "presets load the full pipeline and run probe forward passes — budget a few minutes. "
-                "If the GPU is busy, SVD falls back to the CPU and runs much slower (a WARNING is logged).")
-
-    def _run_extract_krea2(self):
-        """Krea 2: pure weight SVD over all blocks — no pipeline, prompt, or block targeting."""
-        source = self.extract_source_var.get()
-        if not source or not os.path.exists(source):
-            messagebox.showerror("Error", "Please select a valid source LoRA.")
-            return
-
-        output_name = self.extract_output_var.get().strip()
-        if not output_name:
-            messagebox.showerror("Error", "Please enter an output name.")
-            return
-        if not output_name.endswith(".safetensors"):
-            output_name += ".safetensors"
-
-        output_dir = self.prefs_vars["lora_output_dir"].get()
-        os.makedirs(output_dir, exist_ok=True)
-        output_path = os.path.join(output_dir, output_name)
-        # Never silently overwrite (same rule as the Klein path).
-        if os.path.exists(output_path):
-            _stem, _ext = os.path.splitext(output_path)
-            _n = 2
-            while os.path.exists(f"{_stem}_{_n}{_ext}"):
-                _n += 1
-            output_path = f"{_stem}_{_n}{_ext}"
-
-        self.extract_run_btn.configure(state="disabled")
-        self.extract_open_btn.configure(state="disabled")
-        self.extract_log.configure(state="normal")
-        self.extract_log.delete(1.0, tk.END)
-        self.extract_log.configure(state="disabled")
-        self.extract_progress_var.set("Extracting...")
-
-        import threading
-        thread = threading.Thread(
-            target=self._extract_worker_krea2,
-            args=(source, output_path),
-            daemon=True,
-        )
-        thread.start()
-
-    def _extract_worker_krea2(self, source, output_path):
-        """Background worker: weight-only SVD, model-agnostic (extract_weight_only with all blocks)."""
-        try:
-            import sys
-            sys.path.insert(0, os.path.join(os.path.dirname(__file__), "src"))
-            from fizgig.extraction.extractor import LoRAExtractor, ExtractionConfig
-
-            rank = int(self.extract_rank_var.get())
-            config = ExtractionConfig(
-                source_lora_path=source,
-                output_lora_path=output_path,
-                target_rank=rank,
-                timestep_range=(0.0, 1.0),
-                include_blocks=["all"],
-                custom_blocks=None,
-                num_samples=0,
-                prompt="",
-                width=1024,
-                height=1024,
-                seed=42,
-            )
-
-            def progress(stage, current, total):
-                self.master.after(0, lambda: self.extract_progress_var.set(f"{stage}: {current+1}/{total}"))
-
-            _fam_label = ("MiniMax H3" if str(self.extract_family_var.get()) == "minimax"
-                          else "Krea 2")
-            self.master.after(0, lambda: self._extract_log(
-                f"{_fam_label} weight-only SVD (all blocks), rank={rank}\n"))
-            result = LoRAExtractor.extract_weight_only(config, progress_callback=progress)
-
-            # Empty artifact = failure, not success (same rule as the Klein worker).
-            if result.num_layers_extracted == 0:
-                try:
-                    os.remove(result.output_path)
-                except OSError:
-                    pass
-                raise RuntimeError("Extraction produced 0 layers — the source file contains "
-                                   "no LoRA modules this extractor recognises.")
-
-            summary = (f"\nExtraction complete!\n"
-                       f"  Output: {result.output_path}\n"
-                       f"  Layers extracted: {result.num_layers_extracted}\n"
-                       f"  Target rank: {result.target_rank}\n"
-                       f"  Total params: {result.total_params:,}\n"
-                       f"  Time: {result.elapsed_seconds:.1f}s\n")
-            self._extract_output_path = output_path
-
-            def _update_ui():
-                self._extract_log(summary)
-                self.extract_progress_var.set("Done!")
-                self.extract_run_btn.configure(state="normal")
-                self.extract_open_btn.configure(state="normal")
-            self.master.after(0, _update_ui)
-
-        except Exception:
-            import traceback
-            error_msg = f"Extraction failed:\n{traceback.format_exc()}"
-            def _show_error():
-                self._extract_log(error_msg)
-                self.extract_progress_var.set("Error")
-                self.extract_run_btn.configure(state="normal")
-            self.master.after(0, _show_error)
-
-    # endregion
-
-    # region Metadata Tab
-
-    def create_metadata_tab(self):
-        """Create the Metadata tab — view and edit the modelspec metadata on any .safetensors
-        file: LoRAs, DiTs, text encoders, embeddings, whatever — including ones Fizgig didn't
-        train, or trained before these fields existed."""
-        scrollable_frame, _ = self.create_scrollable_frame(self.metadata_tab)
-
-        outer = tk.Frame(scrollable_frame, bg=COLORS["bg_deep"])
-        outer.pack(fill=tk.BOTH, expand=True)
-
-        self._add_tab_banner(
-            outer,
-            "Metadata",
-            "View and edit the SAI ModelSpec metadata embedded in a .safetensors file — title, "
-            "author, description, trigger phrase, thumbnail, and anything else ComfyUI's model "
-            "browser reads. Works on any .safetensors file — LoRA, DiT, text encoder, "
-            "embedding — not just ones Fizgig trained.",
-        )
-
-        self._metadata_editor_path = None
-        self._metadata_editor_thumbnail_uri = None  # current thumbnail data URI, or None
-        self._metadata_editor_custom = {}  # every key outside the standard fields below
-        self._metadata_thumb_photo = None  # keeps the PhotoImage alive; Tk drops it otherwise
-
-        # --- Load ---
-        load_card = self._start_section_card(
-            outer, "File",
-            "Pick any .safetensors file — LoRA, DiT, text encoder, embedding — its current "
-            "metadata loads below.",
-        )
-        load_card.grid_columnconfigure(1, weight=1)
-        ttk.Label(load_card, text="File:").grid(row=0, column=0, sticky=tk.W, padx=(0, 10), pady=4)
-        self.metadata_file_var = tk.StringVar()
-        ttk.Entry(load_card, textvariable=self.metadata_file_var, width=60).grid(
-            row=0, column=1, sticky=tk.EW, pady=4)
-        ttk.Button(load_card, text="Browse", command=self._browse_metadata_file).grid(
-            row=0, column=2, sticky=tk.W, padx=(8, 0), pady=4)
-        self._metadata_status_label = tk.Label(
-            load_card, text="No file loaded.", font=(FONT_FAMILY, 9, "italic"),
-            fg=COLORS["text_muted"], bg=COLORS["bg_surface"])
-        self._metadata_status_label.grid(row=1, column=1, sticky=tk.W, pady=(2, 0))
-
-        # --- Standard fields ---
-        fields_card = self._start_section_card(
-            outer, "Standard Fields",
-            "The fields ComfyUI's model browser (and other spec-aware tools) render specially.",
-        )
-        fields_card.grid_columnconfigure(1, weight=1)
-
-        def _field_row(label_text, row):
-            ttk.Label(fields_card, text=label_text).grid(
-                row=row, column=0, sticky=tk.NW, padx=(0, 10), pady=4)
-            var = tk.StringVar()
-            ttk.Entry(fields_card, textvariable=var, width=60).grid(
-                row=row, column=1, sticky=tk.EW, pady=4)
-            return var
-
-        self.metadata_title_var = _field_row("Title:", 0)
-        self.metadata_author_var = _field_row("Author:", 1)
-        self.metadata_license_var = _field_row("License:", 2)
-        self.metadata_tags_var = _field_row("Tags:", 3)
-        self.metadata_trigger_var = _field_row("Trigger Phrase:", 4)
-        self.metadata_usage_hint_var = _field_row("Usage Hint:", 5)
-
-        ttk.Label(fields_card, text="Description:").grid(
-            row=6, column=0, sticky=tk.NW, padx=(0, 10), pady=4)
-        self.metadata_description_text = tk.Text(
-            fields_card, width=60, height=5, wrap=tk.WORD,
-            bg=COLORS["bg_surface"], fg=COLORS["text_primary"],
-            insertbackground=COLORS["text_primary"], font=(FONT_FAMILY, 10),
-            relief="flat", highlightthickness=1,
-            highlightbackground=COLORS["border"], highlightcolor=COLORS["border_focus"],
-        )
-        self.metadata_description_text.grid(row=6, column=1, sticky=tk.EW, pady=4)
-
-        # --- Thumbnail ---
-        thumb_card = self._start_section_card(
-            outer, "Thumbnail",
-            "The image ComfyUI shows as card art. Fizgig auto-embeds the latest training sample "
-            "when it trains a LoRA — replace or clear it here for any file.",
-        )
-        self._metadata_thumb_label = tk.Label(thumb_card, bg=COLORS["bg_surface"],
-                                              text="(no thumbnail)", fg=COLORS["text_muted"])
-        self._metadata_thumb_label.pack(anchor=tk.W, pady=(0, 8))
-        thumb_btn_row = tk.Frame(thumb_card, bg=COLORS["bg_surface"])
-        thumb_btn_row.pack(anchor=tk.W)
-        ttk.Button(thumb_btn_row, text="Replace...",
-                   command=self._browse_metadata_thumbnail).pack(side=tk.LEFT, padx=(0, 8))
-        ttk.Button(thumb_btn_row, text="Clear",
-                   command=self._clear_metadata_thumbnail).pack(side=tk.LEFT)
-
-        # --- Custom fields ---
-        custom_card = self._start_section_card(
-            outer, "Custom Fields",
-            "Like ID3 tags on an MP3 — the format isn't limited to a fixed list, and a reader "
-            "just ignores whatever it doesn't recognize. Add anything you want: author_email, "
-            "a colorspace profile note, whatever's useful to you. Not part of the SAI ModelSpec "
-            "standard, so tools other than Fizgig won't render these specially, but they're "
-            "stored in the file like any other metadata. Also shows any non-standard keys "
-            "already in the file — nothing gets silently dropped on save.",
-        )
-        tree_frame = tk.Frame(custom_card, bg=COLORS["bg_surface"])
-        tree_frame.pack(fill=tk.BOTH, expand=True)
-        self.metadata_custom_tree = ttk.Treeview(
-            tree_frame, columns=("key", "value"), show="headings", height=6)
-        self.metadata_custom_tree.heading("key", text="Key")
-        self.metadata_custom_tree.heading("value", text="Value")
-        self.metadata_custom_tree.column("key", width=220, anchor=tk.W)
-        self.metadata_custom_tree.column("value", width=420, anchor=tk.W)
-        self.metadata_custom_tree.pack(side=tk.LEFT, fill=tk.BOTH, expand=True)
-        tree_scroll = ttk.Scrollbar(tree_frame, orient=tk.VERTICAL,
-                                    command=self.metadata_custom_tree.yview)
-        tree_scroll.pack(side=tk.LEFT, fill=tk.Y)
-        self.metadata_custom_tree.configure(yscrollcommand=tree_scroll.set)
-
-        custom_btn_row = tk.Frame(custom_card, bg=COLORS["bg_surface"])
-        custom_btn_row.pack(anchor=tk.W, pady=(8, 0))
-        ttk.Button(custom_btn_row, text="Add field...",
-                   command=self._add_metadata_custom_field).pack(side=tk.LEFT, padx=(0, 8))
-        ttk.Button(custom_btn_row, text="Remove selected",
-                   command=self._remove_metadata_custom_field).pack(side=tk.LEFT)
-
-        # --- Save ---
-        save_row = tk.Frame(outer, bg=COLORS["bg_deep"])
-        save_row.pack(fill=tk.X, padx=36, pady=(0, 20))
-        ttk.Button(save_row, text="Save",
-                   command=lambda: self._save_metadata_file(save_as=False)).pack(
-            side=tk.LEFT, padx=(0, 8))
-        ttk.Button(save_row, text="Save As...",
-                   command=lambda: self._save_metadata_file(save_as=True)).pack(side=tk.LEFT)
-        self._metadata_save_status = tk.Label(
-            save_row, text="", font=(FONT_FAMILY, 9, "italic"),
-            fg=COLORS["text_muted"], bg=COLORS["bg_deep"])
-        self._metadata_save_status.pack(side=tk.LEFT, padx=(16, 0))
-
-    def _browse_metadata_file(self):
-        filepath = filedialog.askopenfilename(
-            title="Select a .safetensors file",
-            filetypes=[("SafeTensors", "*.safetensors"), ("All files", "*.*")],
-            initialdir=self._lora_initialdir(),
-        )
-        if filepath:
-            self.metadata_file_var.set(filepath)
-            self._load_metadata_file(filepath)
-
-    def _load_metadata_file(self, path):
-        from fizgig.training.metadata import load_metadata_from_safetensors
-        try:
-            meta = load_metadata_from_safetensors(path)
-        except Exception as e:
-            messagebox.showerror("Error", f"Could not read metadata:\n{e}")
-            return
-
-        self._metadata_editor_path = path
-        standard = {
-            "modelspec.title": self.metadata_title_var,
-            "modelspec.author": self.metadata_author_var,
-            "modelspec.license": self.metadata_license_var,
-            "modelspec.tags": self.metadata_tags_var,
-            "modelspec.trigger_phrase": self.metadata_trigger_var,
-            "modelspec.usage_hint": self.metadata_usage_hint_var,
-        }
-        for key, var in standard.items():
-            var.set(meta.get(key, "") or "")
-
-        self.metadata_description_text.delete("1.0", tk.END)
-        self.metadata_description_text.insert("1.0", meta.get("modelspec.description", "") or "")
-
-        self._metadata_editor_thumbnail_uri = meta.get("modelspec.thumbnail")
-        self._show_metadata_thumbnail_preview(self._metadata_editor_thumbnail_uri)
-
-        skip_keys = set(standard.keys()) | {"modelspec.description", "modelspec.thumbnail"}
-        self._metadata_editor_custom = {k: v for k, v in meta.items() if k not in skip_keys}
-        self._refresh_metadata_custom_tree()
-
-        n = len(meta)
-        self._metadata_status_label.config(
-            text=f"Loaded — {n} metadata key{'s' if n != 1 else ''} found.",
-            fg=COLORS["text_secondary"])
-        self._metadata_save_status.config(text="")
-
-    def _show_metadata_thumbnail_preview(self, data_uri):
-        if not data_uri or not str(data_uri).startswith("data:image"):
-            self._metadata_thumb_label.config(image="", text="(no thumbnail)")
-            self._metadata_thumb_photo = None
-            return
-        try:
-            import base64
-            from io import BytesIO
-            b64 = data_uri.split(",", 1)[1]
-            img = Image.open(BytesIO(base64.b64decode(b64)))
-            img.thumbnail((256, 256))
-            photo = ImageTk.PhotoImage(img)
-            self._metadata_thumb_photo = photo  # reference kept alive deliberately
-            self._metadata_thumb_label.config(image=photo, text="")
-        except Exception:
-            self._metadata_thumb_label.config(image="", text="(couldn't decode thumbnail)")
-            self._metadata_thumb_photo = None
-
-    def _browse_metadata_thumbnail(self):
-        filepath = filedialog.askopenfilename(
-            title="Select a thumbnail image",
-            filetypes=[("Images", "*.png *.jpg *.jpeg *.webp"), ("All files", "*.*")],
-        )
-        if not filepath:
-            return
-        from fizgig.training.metadata import thumbnail_data_uri
-        uri = thumbnail_data_uri(filepath)
-        if not uri:
-            messagebox.showerror("Error", "Could not read that image.")
-            return
-        self._metadata_editor_thumbnail_uri = uri
-        self._show_metadata_thumbnail_preview(uri)
-
-    def _clear_metadata_thumbnail(self):
-        self._metadata_editor_thumbnail_uri = None
-        self._show_metadata_thumbnail_preview(None)
-
-    def _refresh_metadata_custom_tree(self):
-        self.metadata_custom_tree.delete(*self.metadata_custom_tree.get_children())
-        for k, v in sorted(self._metadata_editor_custom.items()):
-            display_v = v if len(str(v)) <= 120 else str(v)[:117] + "..."
-            self.metadata_custom_tree.insert("", tk.END, iid=k, values=(k, display_v))
-
-    def _add_metadata_custom_field(self):
-        dlg = tk.Toplevel(self.master)
-        dlg.title("Add custom field")
-        dlg.configure(bg=BG_COLOR)
-        dlg.transient(self.master)
-        tk.Label(dlg, text="Key:", bg=BG_COLOR, fg=COLORS["text_secondary"]).grid(
-            row=0, column=0, sticky=tk.W, padx=10, pady=(10, 2))
-        key_entry = ttk.Entry(dlg, width=40)
-        key_entry.grid(row=0, column=1, padx=10, pady=(10, 2))
-        tk.Label(dlg, text="Value:", bg=BG_COLOR, fg=COLORS["text_secondary"]).grid(
-            row=1, column=0, sticky=tk.W, padx=10, pady=2)
-        val_entry = ttk.Entry(dlg, width=40)
-        val_entry.grid(row=1, column=1, padx=10, pady=2)
-
-        def ok():
-            k = key_entry.get().strip()
-            v = val_entry.get().strip()
-            if k:
-                self._metadata_editor_custom[k] = v
-                self._refresh_metadata_custom_tree()
-            dlg.destroy()
-
-        btn_row = tk.Frame(dlg, bg=BG_COLOR)
-        btn_row.grid(row=2, column=0, columnspan=2, pady=10)
-        ttk.Button(btn_row, text="Cancel", command=dlg.destroy).pack(side=tk.LEFT, padx=5)
-        ttk.Button(btn_row, text="Add", command=ok).pack(side=tk.LEFT, padx=5)
-        key_entry.bind("<Return>", lambda e: ok())
-        key_entry.focus_set()
-        dlg.grab_set()
-
-    def _remove_metadata_custom_field(self):
-        for k in self.metadata_custom_tree.selection():
-            self._metadata_editor_custom.pop(k, None)
-        self._refresh_metadata_custom_tree()
-
-    def _save_metadata_file(self, save_as=False):
-        if not self._metadata_editor_path:
-            messagebox.showinfo("No file loaded", "Load a .safetensors file first.")
-            return
-
-        dest = self._metadata_editor_path
-        if save_as:
-            dest = filedialog.asksaveasfilename(
-                title="Save file as",
-                defaultextension=".safetensors",
-                filetypes=[("SafeTensors", "*.safetensors")],
-                initialfile=os.path.basename(self._metadata_editor_path),
-            )
-            if not dest:
-                return
-        elif not messagebox.askyesno(
-                "Overwrite?",
-                f"Save metadata changes to:\n{dest}\n\nThis overwrites the file in place."):
-            return
-
-        try:
-            from safetensors.torch import load_file, save_file
-            # .clone() forces a real copy out of the mmap'd view load_file returns. Without
-            # this, saving back onto the SAME path fails on Windows with "the requested
-            # operation cannot be performed on a file with a user-mapped section open"
-            # (error 1224) — the mapping from this exact load is still active, and Windows
-            # (unlike Linux) refuses to overwrite a file while it's mapped. Cloning breaks
-            # that dependency before we ever try to write.
-            tensors = {k: v.clone() for k, v in load_file(self._metadata_editor_path).items()}
-        except Exception as e:
-            messagebox.showerror("Error", f"Could not read the file's tensors:\n{e}")
-            return
-
-        new_meta = dict(self._metadata_editor_custom)
-        field_map = {
-            "modelspec.title": self.metadata_title_var.get().strip(),
-            "modelspec.author": self.metadata_author_var.get().strip(),
-            "modelspec.license": self.metadata_license_var.get().strip(),
-            "modelspec.tags": self.metadata_tags_var.get().strip(),
-            "modelspec.trigger_phrase": self.metadata_trigger_var.get().strip(),
-            "modelspec.usage_hint": self.metadata_usage_hint_var.get().strip(),
-            "modelspec.description": self.metadata_description_text.get("1.0", "end-1c").strip(),
-        }
-        for k, v in field_map.items():
-            if v:
-                new_meta[k] = v
-        if self._metadata_editor_thumbnail_uri:
-            new_meta["modelspec.thumbnail"] = self._metadata_editor_thumbnail_uri
-
-        # A metadata-only edit still goes through a full resave, which isn't guaranteed to be
-        # byte-identical to the original — so hashes computed over the old bytes can no longer
-        # be trusted. Same move bake.py already makes whenever it changes a file's contents.
-        for stale in ("sshs_model_hash", "sshs_legacy_hash", "modelspec.hash_sha256"):
-            new_meta.pop(stale, None)
-
-        # Write to a temp file and swap it in atomically, so a failed/interrupted save can
-        # never leave the original file half-written.
-        tmp_dest = dest + ".tmp"
-        try:
-            save_file(tensors, tmp_dest, metadata=new_meta)
-            os.replace(tmp_dest, dest)
-        except Exception as e:
-            try:
-                if os.path.exists(tmp_dest):
-                    os.remove(tmp_dest)
-            except OSError:
-                pass
-            messagebox.showerror("Error", f"Could not save:\n{e}")
-            return
-
-        self._metadata_save_status.config(text=f"Saved {os.path.basename(dest)}",
-                                          fg=COLORS["text_secondary"])
-        if save_as:
-            self.metadata_file_var.set(dest)
-            self._metadata_editor_path = dest
-
-    # endregion
-
-    # region Preferences Tab
-
-    # The paths a family cannot train without. The optional rows (reference DiT, audio VAE,
-    # Turbo LoRA) neither hold a section open nor count against its badge — a family whose only
-    # gaps are optional is configured.
-    _PREFS_FAMILY_KEYS = {
-        "klein": ("base_dit", "distilled_dit", "vae", "text_encoder"),
-        "krea2": ("krea2_raw_dit", "krea2_turbo_dit", "krea2_vae", "krea2_text_encoder"),
-        "minimax": ("minimax_dit", "minimax_text_encoder", "minimax_vae"),
-    }
-    # Badge names — "2 paths needed if training Klein 9B" reads as advice, where a bare
-    # "2 paths needed" reads as a problem: someone who never trains that family owes it nothing.
-    _PREFS_FAMILY_NAMES = {"klein": "Klein 9B", "krea2": "Krea 2", "minimax": "MiniMax H3"}
-
-    def _prefs_family_section(self, parent, family, title, description):
-        """A collapsible model-path section for one model family on the Preferences tab.
-
-        Starts collapsed only when the family's required paths are all filled: a new user sent
-        here by the Start tab's setup prompt lands on the sections they still need already open,
-        while a configured machine shows three closed headers instead of a wall of path rows.
-        The header badge keeps a collapsed section honest — you can see whether a family needs
-        attention without opening it. Returns (content_frame, first_free_grid_row).
-        """
-        keys = self._PREFS_FAMILY_KEYS[family]
-
-        def _missing():
-            return sum(1 for k in keys if not str(self.prefs_vars[k].get() or "").strip())
-
-        section = CollapsibleFrame(parent, title, default_expanded=_missing() > 0)
-        section.pack(fill=tk.X, padx=36, pady=(0, 16))
-        content = section.get_content_frame()
-        content.columnconfigure(1, weight=1)
-        tk.Label(content, text=description, font=(FONT_FAMILY, 10),
-                 fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                 wraplength=760, justify=tk.LEFT
-                 ).grid(row=0, column=0, columnspan=3, sticky=tk.W, pady=(4, 10))
-
-        def _refresh_badge(*_a):
-            n = _missing()
-            try:
-                section.badge.config(
-                    text="✓ configured" if n == 0
-                    else (f"{n} path{'s' if n != 1 else ''} needed if training "
-                          f"{self._PREFS_FAMILY_NAMES[family]}"),
-                    fg=COLORS["text_secondary"] if n == 0 else COLORS["warning"])
-            except tk.TclError:
-                pass
-
-        _refresh_badge()
-        for k in keys:
-            self.prefs_vars[k].trace_add("write", _refresh_badge)
-        return content, 1
-
-    def create_prefs_tab(self):
-        """Create the Preferences tab (Start-tab styled)."""
-        scrollable_frame, _ = self.create_scrollable_frame(self.prefs_tab)
-
-        outer = tk.Frame(scrollable_frame, bg=COLORS["bg_deep"])
-        outer.pack(fill=tk.BOTH, expand=True)
-
-        self._add_tab_banner(
-            outer,
-            "Preferences",
-            "Centralised paths + inference performance knobs. Changes here propagate to every tab automatically "
-            "and persist to prefs.json.",
-        )
-
-        # The three model-family sections are collapsible, and smart about it: a family with a
-        # required path still blank starts open, a configured one starts closed. Click the
-        # header to toggle; the badge says which state you're in without opening anything.
-        models_card, next_row = self._prefs_family_section(
-            outer, "klein", "Model Paths (Klein 9B)",
-            "Absolute paths to the four model files. Each row has a Download link that opens the HuggingFace page "
-            "in your browser.",
-        )
-        next_row = self._add_pref_row(
-            models_card, next_row, "Base DiT:", "base_dit",
-            "Klein 9B Base model (for training & precise profiling). "
-            "Recommended: the fp8 version — same training quality, ~half the VRAM (stays resident at "
-            "~9.6GB, fits 16GB cards). The bf16 version is the larger alternative.",
-            download_url="https://huggingface.co/black-forest-labs/FLUX.2-klein-base-9b-fp8/tree/main",
-            download_label="Download fp8 (recommended)",
-            download_note="~9.5GB fp8 — Black Forest Labs (flux-2-klein-base-9b-fp8.safetensors)",
-            download_url2="https://huggingface.co/black-forest-labs/FLUX.2-klein-base-9B/tree/main",
-            download_label2="Download bf16",
-            download_note2="~17GB bf16 (flux-2-klein-base-9b.safetensors)",
-        )
-        next_row = self._add_pref_row(
-            models_card, next_row, "Distilled DiT:", "distilled_dit",
-            "Klein 9B Distilled model (for Repair Studio previews, fast profiling & diagnostics)",
-            download_url="https://huggingface.co/black-forest-labs/FLUX.2-klein-9b-fp8/tree/main",
-            download_note="~9GB fp8 quantised — Black Forest Labs (flux-2-klein-9b-fp8.safetensors)",
-        )
-        next_row = self._add_pref_row(
-            models_card, next_row, "VAE / AE:", "vae",
-            "Flux 2 AutoEncoder — use ae.safetensors from FLUX.2-dev root (NOT the vae/ subfolder Diffusers file)",
-            download_url="https://huggingface.co/black-forest-labs/FLUX.2-dev/blob/main/ae.safetensors",
-            download_note="~320MB  ·  get ae.safetensors from FLUX.2-dev root  ·  NOT vae/diffusion_pytorch_model.safetensors (Diffusers format, incompatible)",
-        )
-        next_row = self._add_pref_row(
-            models_card, next_row, "Text Encoder:", "text_encoder",
-            "Qwen3-8B text encoder (used by Klein 9B)",
-            download_url="https://huggingface.co/Comfy-Org/vae-text-encorder-for-flux-klein-9b/blob/main/split_files/text_encoders/qwen_3_8b.safetensors",
-            download_note="~15GB single-file safetensors — Qwen3-8B packaged for Klein 9B (Comfy-Org)",
-        )
-        # Klein users routinely skip the whole Krea 2 card, and would never learn that one file
-        # in it is a better captioner than Florence-2 for THEIR datasets — captions are just
-        # .txt, so the model family is irrelevant. Worth saying here, where they are already
-        # filling in paths, rather than hoping they read the card below.
-        _qwen_tip = tk.Label(
-            models_card,
-            text="💡 Training Klein only? The Krea 2 Qwen3-VL text encoder is still worth having — "
-                 "the Captions tab can caption ANY dataset with it, following an instruction you "
-                 "can edit, and it writes better training captions than Florence-2. The download "
-                 "button below fetches it for you; nothing else about Krea 2 is needed.",
-            font=(FONT_FAMILY, 9), fg=COLORS["text_secondary"], bg=COLORS["bg_surface"],
-            wraplength=760, justify=tk.LEFT)
-        _qwen_tip.grid(row=next_row, column=0, columnspan=3, sticky=tk.W, pady=(12, 2))
-        next_row += 1
-        self._add_fetch_models_row(
-            models_card, next_row, "klein",
-            "Fetches the four files above plus the Krea 2 Qwen3-VL captioning text encoder "
-            "(~39 GB all in) and fills in these paths for you, plus the small helper models "
-            "(Florence-2 captioner, face model for the Look Filter and likeness scoring, EN→ZH "
-            "translator, Gizmo's Whisper transcriber — ~1.9 GB) so nothing stalls to download "
-            "later and everything works offline. "
-            "Black Forest Labs gate their downloads, so you'll need a free HuggingFace token — "
-            "Fizgig asks for it and tells you which pages to accept the licence on.")
-        next_row += 1
-
-        # Krea 2 model paths
-        krea_card, kr = self._prefs_family_section(
-            outer, "krea2", "Model Paths (Krea 2)",
-            "Krea 2 LoRA training + inference. Train on RAW; previews and inference use the pre-quant fp8 Turbo "
-            "(8-step, CFG-free). The text encoder can be either Qwen3-VL-4B file — bf16, or the smaller "
-            "fp8_scaled (~3.6 GB less VRAM; its vision tower is bf16 either way).",
-        )
-        kr = self._add_pref_row(
-            krea_card, kr, "RAW DiT:", "krea2_raw_dit",
-            "Krea 2 RAW (undistilled 12.9B base) — the training model (krea2_raw_bf16.safetensors)",
-            download_url="https://huggingface.co/Comfy-Org/Krea-2/blob/main/diffusion_models/krea2_raw_bf16.safetensors",
-            download_note="~26GB bf16 — Comfy-Org/Krea-2 → diffusion_models/krea2_raw_bf16.safetensors (train on this)",
-        )
-        kr = self._add_pref_row(
-            krea_card, kr, "Turbo DiT (fp8):", "krea2_turbo_dit",
-            "Krea 2 Turbo, pre-quantized fp8 (ComfyUI) — fast previews + inference (8-step, CFG-free)",
-            download_url="https://huggingface.co/Comfy-Org/Krea-2/blob/main/diffusion_models/krea2_turbo_fp8_scaled.safetensors",
-            download_note="~13GB fp8 — Comfy-Org/Krea-2 → diffusion_models/krea2_turbo_fp8_scaled.safetensors",
-        )
-        kr = self._add_pref_row(
-            krea_card, kr, "Qwen-Image VAE:", "krea2_vae",
-            "The Qwen-Image VAE used by Krea 2 (qwen_image_vae.safetensors)",
-            download_url="https://huggingface.co/Comfy-Org/Krea-2/blob/main/vae/qwen_image_vae.safetensors",
-            download_note="~250MB — Comfy-Org/Krea-2 → vae/qwen_image_vae.safetensors",
-        )
-        kr = self._add_pref_row(
-            krea_card, kr, "Qwen3-VL TE:", "krea2_text_encoder",
-            "Qwen3-VL-4B text encoder. fp8_scaled is recommended: it quantises only the language "
-            "layers and ships the same full bf16 vision tower, so reference images and AI "
-            "captioning are unaffected — measured 4.9 GB resident vs 8.3 GB, for captions we "
-            "could not tell apart",
-            download_url="https://huggingface.co/Comfy-Org/Krea-2/blob/main/text_encoders/qwen3vl_4b_fp8_scaled.safetensors",
-            download_label="Download fp8 (recommended)",
-            download_note="~5.2GB fp8_scaled — 3.4 GB less VRAM than bf16, virtually identical output",
-            download_url2="https://huggingface.co/Comfy-Org/Krea-2/blob/main/text_encoders/qwen3vl_4b_bf16.safetensors",
-            download_label2="bf16",
-            download_note2="~8.9GB bf16 — the full-precision original, if you would rather not "
-                           "quantise at all",
-        )
-        kr = self._add_pref_row(
-            krea_card, kr, "Turbo LoRA (rank 64):", "krea2_turbo_lora",
-            "Turbo distillation as a LoRA — RAW + this at strength 1.0 samples like the Turbo model "
-            "(same 8-step CFG-free settings), letting previews run on the training DiT without loading "
-            "the separate Turbo checkpoint",
-            download_url="https://huggingface.co/Comfy-Org/Krea-2/blob/main/loras/krea2_turbo_lora_rank_64_bf16.safetensors",
-            download_note="~470MB bf16 — Comfy-Org/Krea-2 → loras/krea2_turbo_lora_rank_64_bf16.safetensors",
-        )
-        self._add_fetch_models_row(
-            krea_card, kr + 1, "krea2",
-            "Fetches every file above (~45 GB) and fills in these paths for you, plus the small "
-            "helper models (Florence-2 captioner, face model for the Look Filter and likeness "
-            "scoring, EN→ZH translator, Gizmo's Whisper transcriber — ~1.9 GB) so nothing "
-            "stalls to download later and everything works offline. No HuggingFace account "
-            "needed — none of these are gated.")
-        _offline_tip = tk.Label(
-            krea_card,
-            text="💡 Already have these files for ComfyUI? Filling the paths in by hand works "
-                 "perfectly — the download button is a convenience, not a requirement. The first "
-                 "time you caption or train while online, Fizgig quietly fetches a few tiny "
-                 "helper files and keeps them, and from then on everything runs fully offline. "
-                 "Setting up a machine that will never see the internet? Paste a complete "
-                 "HuggingFace model folder into the text encoder field instead of a single file "
-                 "and nothing needs downloading at all.",
-            font=(FONT_FAMILY, 9), fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-            wraplength=760, justify=tk.LEFT)
-        _offline_tip.grid(row=kr + 2, column=0, columnspan=3, sticky=tk.W, pady=(12, 2))
-
-        # MiniMax H3 model paths — third family, beside the other two now that clip+audio
-        # training has outgrown its bottom-of-the-page beginnings.
-        mm_card, mr = self._prefs_family_section(
-            outer, "minimax", "Model Paths (MiniMax H3 — experimental)",
-            "Image-only LoRA training for MiniMax's ~33B H3 omni DiT. Train on the pruned int8 DiT "
-            "— the same file ComfyUI runs — quantized to NF4 at load, so the resident base is "
-            "~11 GB. The Qwen3-VL-32B text encoder and the video VAE are only needed for the "
-            "one-time caching pass; the compact nvfp4 TE is recommended. Trains on stills, or on "
-            "short video clips — and with the audio VAE set, on their sound too.",
-        )
-        mr = self._add_pref_row(
-            mm_card, mr, "DiT:", "minimax_dit",
-            "MiniMax H3 DiT — the training base. Use the PRUNED int8 file "
-            "(minimax_h3_fl2va_pruned_int8_convrot.safetensors, ~21 GB): it is the one ComfyUI "
-            "runs, so your LoRA trains against the weights it will be deployed on, and its "
-            "curve-table AdaLN is a target a LoRA can actually use. The ~66 GB bf16 file also "
-            "works. The pruned file KEEPS its int8 weights (~21 GB on the GPU, what the reference "
-            "trainer does); the bf16 file is quantized to NF4 at load (~11 GB, a little lossier).",
-            download_url="https://huggingface.co/Comfy-Org/MiniMax-H3/blob/main/diffusion_models/minimax_h3_fl2va_pruned_int8_convrot.safetensors",
-            download_note="~21GB — Comfy-Org/MiniMax-H3 → diffusion_models/minimax_h3_fl2va_pruned_int8_convrot.safetensors (fl2va is the trainable variant; the 66GB bf16 file works too)",
-        )
-        mr = self._add_pref_row(
-            mm_card, mr, "DiT (reference):", "minimax_ref_dit",
-            "OPTIONAL — used when the Training tab's Training Base is set to Reference "
-            "(ref2va), and for reference distillation ('Learn identity from'). This is the "
-            "ref2va model, a DIFFERENT fine-tune from the fl2va one above and not just another "
-            "quantization of it: it is what ComfyUI's Reference-to-Video workflow loads, and "
-            "the only H3 build that accepts reference images. A LoRA trained on it is most "
-            "faithful deployed on it. Leave blank if you only train on the standard base.",
-            download_url="https://huggingface.co/Comfy-Org/MiniMax-H3/blob/main/diffusion_models/minimax_h3_ref2va_pruned_int8_convrot.safetensors",
-            download_note="~21GB — Comfy-Org/MiniMax-H3 -> diffusion_models/"
-                          "minimax_h3_ref2va_pruned_int8_convrot.safetensors (the pruned int8 "
-                          "build, same shape as the fl2va one above; you may already have it if "
-                          "you use the r2v workflow)",
-        )
-        mr = self._add_pref_row(
-            mm_card, mr, "Qwen3-VL-32B TE:", "minimax_text_encoder",
-            "Qwen3-VL-32B text encoder — nvfp4 (the compact ComfyUI file) or bf16 both work; the "
-            "loader detects which you gave it. The nvfp4 file keeps its packed weights (~15.7 GB "
-            "on the GPU); bf16 is NF4-quantized at load (~14 GB). Used only while caching caption "
-            "embeddings, then offloaded before training. (The int8_convrot TE "
-            "variant is NOT supported — its rotated weights can't be dequantized here.)",
-            download_url="https://huggingface.co/Comfy-Org/MiniMax-H3/blob/main/text_encoders/qwen3vl_32b_minimax_h3_nvfp4_awq.safetensors",
-            download_label="Download nvfp4 (recommended)",
-            download_note="~15.7GB nvfp4-awq — the same TE ComfyUI uses, so you may already have it; "
-                          "identical conditioning to bf16 (validated), just a slower one-off load",
-            download_url2="https://huggingface.co/Comfy-Org/MiniMax-H3/blob/main/text_encoders/qwen3vl_32b_minimax_h3_bf16.safetensors",
-            download_label2="bf16",
-            download_note2="~51.5GB bf16 — the full-precision original; loads faster, 3.3x the disk",
-        )
-        mr = self._add_pref_row(
-            mm_card, mr, "Video VAE:", "minimax_vae",
-            "The H3 video VAE — encodes each training image to a 24-channel latent (used only "
-            "during caching).",
-            download_url="https://huggingface.co/Comfy-Org/MiniMax-H3/blob/main/vae/minimax_h3_video_vae_fp16.safetensors",
-            download_note="~4.9GB — Comfy-Org/MiniMax-H3 → vae/minimax_h3_video_vae_fp16.safetensors",
-        )
-        mr = self._add_pref_row(
-            mm_card, mr, "Audio VAE:", "minimax_audio_vae",
-            "OPTIONAL — set this to train on the sound in your video clips. H3 generates audio and "
-            "video together, so a clip with sound can teach it a voice, and nothing else can. "
-            "Used only during caching, and only by clips: a folder of stills never loads it, and "
-            "neither does a clip you muted (a _mute on the filename trains that clip's video and "
-            "ignores its sound). Leave blank and clips train silent, exactly as they did before.",
-            download_url="https://huggingface.co/Comfy-Org/MiniMax-H3/blob/main/vae/minimax_h3_audio_vae_fp32.safetensors",
-            download_note="~605MB — Comfy-Org/MiniMax-H3 → vae/minimax_h3_audio_vae_fp32.safetensors",
-        )
-        mr = self._add_pref_row(
-            mm_card, mr, "Turbo LoRA:", "minimax_turbo_lora",
-            "OPTIONAL — fast in-training previews. With this set, previews render in 6 steps "
-            "with the community Turbo LoRA applied at 75% on top of your training LoRA — the "
-            "same pairing fast ComfyUI inference uses — instead of the full 20-step pass. It "
-            "touches PREVIEWS ONLY: the Turbo is switched in for the sample render and out "
-            "again before the next training step, and your saved LoRA never contains it. "
-            "Steps and strength are adjustable on the Samples tab.",
-            download_url="https://huggingface.co/larryvrh/MiniMax-H3-Turbo-Lora/blob/main/minimax_h3_turbo_v4_step600.safetensors",
-            download_note="~780MB — larryvrh/MiniMax-H3-Turbo-Lora → "
-                          "minimax_h3_turbo_v4_step600.safetensors (you may already have it in "
-                          "ComfyUI's loras folder)",
-        )
-        self._add_fetch_models_row(
-            mm_card, mr, "minimax",
-            "Fetches the DiT, text encoder, both VAEs and the Turbo LoRA above, plus the Krea 2 Qwen3-VL captioning "
-            "text encoder (~47 GB all in), and fills in these paths for you — plus the small "
-            "helper models (Florence-2 captioner, face model for the Look "
-            "Filter and likeness scoring, EN→ZH translator, Gizmo's Whisper transcriber — "
-            "~1.9 GB) so nothing stalls to download later and everything works offline. No "
-            "HuggingFace account needed — none of these are gated. The "
-            "reference DiT is left out unless you tick it above: another 21 GB, and it is only "
-            "used by identity mode.",
-            optional_label="Include the reference DiT (+21 GB)")
-
-        # Card 1b: which GPU. Only when the machine actually has more than one - a chooser with a
-        # single entry is noise, and the whole feature is a no-op there.
-        _gpus = _enumerate_gpus()
-        if len(_gpus) > 1:
-            gpu_card = self._start_section_card(
-                outer, "Graphics Card",
-                "Which GPU Fizgig uses — for training and for the workbench tools alike. "
-                "Everything else in the app then treats that card as the only one there is.",
-            )
-            gpu_card.columnconfigure(1, weight=1)
-            ttk.Label(gpu_card, text="Use GPU:").grid(row=0, column=0, sticky=tk.W,
-                                                      padx=(0, 10), pady=4)
-            self._gpu_choice_labels = {"": "System default (GPU 0)"}
-            self._gpu_info = {}
-            for _i, _name, _gb, _uuid in _gpus:
-                self._gpu_choice_labels[str(_i)] = f"{_i}: {_name} ({_gb:.0f} GB)"
-                self._gpu_info[str(_i)] = (_i, _name, _gb, _uuid)
-            self._nvml_init = False  # reset so first _read_vram() uses correct index
-            _saved_gpu = str(self.prefs.get("cuda_device", "")).strip()
-            # Match saved value: UUID (new format) or index (legacy prefs.json)
-            _matched_key = ""
-            if _saved_gpu:
-                _matched_key = next(
-                    (k for k, v in self._gpu_choice_labels.items()
-                     if self._gpu_info.get(k, (None, None, None, None))[3] == _saved_gpu
-                     or k == _saved_gpu),
-                    "")
-            self._gpu_choice_var = tk.StringVar(
-                value=self._gpu_choice_labels.get(
-                    _matched_key,
-                    self._gpu_choice_labels[""]))
-            _gpu_combo = ttk.Combobox(
-                gpu_card, textvariable=self._gpu_choice_var,
-                values=list(self._gpu_choice_labels.values()), width=44, state="readonly")
-            _gpu_combo.grid(row=0, column=1, sticky=tk.W, pady=4)
-            _gpu_combo.bind("<<ComboboxSelected>>", self._on_gpu_choice)
-            _gpu_note = tk.Label(
-                gpu_card,
-                text=("Takes effect for the next training run straight away. The in-app tools "
-                      "(Repair Studio, Explorer, Royale, Profiler, Extract) hold on to the card "
-                      "they started with, so restart Fizgig to move those."),
-                font=(FONT_FAMILY, 9), fg=COLORS["text_muted"], bg=COLORS["bg_surface"],
-                wraplength=720, justify=tk.LEFT)
-            _gpu_note.grid(row=1, column=1, sticky=tk.W, pady=(0, 4))
-            if os.environ.get("CUDA_VISIBLE_DEVICES") and not str(
-                    self.prefs.get("cuda_device", "")).strip():
-                tk.Label(gpu_card,
-                         text=(f"CUDA_VISIBLE_DEVICES={os.environ['CUDA_VISIBLE_DEVICES']} is set "
-                               f"in your environment and wins over this setting."),
-                         font=(FONT_FAMILY, 9), fg=COLORS["warning"],
-                         bg=COLORS["bg_surface"], wraplength=720, justify=tk.LEFT
-                         ).grid(row=2, column=1, sticky=tk.W, pady=(0, 4))
-
-        # Card 2: Inference Performance
-        inf_card = self._start_section_card(
-            outer, "Inference Performance",
-            "DiT Block Swap moves transformer blocks to CPU during forward passes to cut VRAM, at the cost of PCIe "
-            "latency per step. Affects the workbench tools — Repair Studio / Profiler / Extract / Explorer. "
-            "Training (and its Distilled samples) manage their own swap automatically and ignore this setting.",
-        )
-        inf_card.columnconfigure(1, weight=1)
-
-        ttk.Label(inf_card, text="DiT Block Swap (inference):").grid(row=0, column=0, sticky=tk.W, padx=(0, 10), pady=4)
-        # Workbench previews (Repair Studio / Profiler / Extract) default to the
-        # Distilled fp8 model (light); loading Base is heavier. VRAM shown for each.
-        inference_swap_options = [
-            "Auto (detect from GPU)",
-            "0  (Distilled fp8: 16GB / Base: 24GB)",
-            "4  (Distilled fp8: 14GB / Base: 20GB)",
-            "8  (Distilled fp8: 11GB / Base: 16GB)",
-            "12 (Distilled fp8: 10GB / Base: 14GB)",
-            "16 (Distilled fp8: 8GB / Base: 12GB)",
-        ]
-        _inf_combo = ttk.Combobox(
-            inf_card, textvariable=self.prefs_vars["inference_blocks_to_swap"],
-            values=inference_swap_options, width=40, state="readonly",
-        )
-        _inf_combo.grid(row=0, column=1, sticky=tk.W, pady=4)
-        # Snap whatever's saved to the matching new label by extracting the leading integer.
-        import re as _re_snap
-        _current = str(self.prefs_vars["inference_blocks_to_swap"].get()).strip()
-        _m = _re_snap.match(r'\d+', _current)
-        if _m:
-            _leading_int = _m.group()
-            for _opt in inference_swap_options:
-                if _opt.lstrip().startswith(_leading_int + " ") or _opt.lstrip().startswith(_leading_int + "  "):
-                    self.prefs_vars["inference_blocks_to_swap"].set(_opt)
-                    break
-
-        # INT8 fast inference (on by default). Quantizes the workbench/preview DiT's block Linears to
-        # int8 for a faster matmul. Same VRAM as fp8 (8-bit either way) — a speed knob, not a memory
-        # one — stacks with block swap, and only affects previews (never the saved LoRA). Biggest win
-        # on RTX 30-series (no fast fp8), modest on 40/50-series.
-        ttk.Label(inf_card, text="INT8 fast inference:").grid(row=1, column=0, sticky=tk.W, padx=(0, 10), pady=4)
-        ttk.Checkbutton(
-            inf_card, text="Quantize previews/workbench to int8 for a faster matmul (recommended)",
-            variable=self.prefs_vars["inference_int8"], onvalue="1", offvalue="0",
-        ).grid(row=1, column=1, sticky=tk.W, pady=4)
-        tk.Label(inf_card, text="On by default. Same VRAM as fp8, near-identical quality, previews only. Turn off to use fp8.",
-                 font=(FONT_FAMILY, 9), fg=COLORS["text_muted"], bg=COLORS["bg_surface"]).grid(
-            row=2, column=1, sticky=tk.W, padx=5, pady=(0, 4))
-
-        # Card 3: Output Directories
-        out_card = self._start_section_card(
-            outer, "Output Directories",
-            "Paths stored as relative-to-repo when they live inside FizgigIndependent/ (portable across clones/moves), "
-            "absolute otherwise. Dataset TOMLs always live in FizgigIndependent/dataset/ — not configurable.",
-        )
-        out_card.columnconfigure(1, weight=1)
-        next_row = 0
-        next_row = self._add_pref_row(out_card, next_row, "LoRA output:", "lora_output_dir", "Where trained LoRAs are saved", is_dir=True)
-        next_row = self._add_pref_row(out_card, next_row, "Profiles:", "profiles_dir", "Where profiler HTML reports are saved", is_dir=True)
-        next_row = self._add_pref_row(out_card, next_row, "Cache:", "cache_dir", "Cached latents and text encodings", is_dir=True)
-
-        # Card 3b: Input Directories — default starting folders for the Browse
-        # dialogs, so users don't re-hunt for their LoRAs / reference images.
-        in_card = self._start_section_card(
-            outer, "Input Directories",
-            "Default folders the Browse dialogs open in. Saves hunting each session when loading "
-            "LoRAs or reference images. Leave empty to use the last-used folder.",
-        )
-        in_card.columnconfigure(1, weight=1)
-        in_row = 0
-        in_row = self._add_pref_row(in_card, in_row, "LoRA loading:", "input_lora_dir",
-                                    "Default folder for loading LoRAs — Repair Studio, LoRA the Explorer, and the Context LoRA picker", is_dir=True)
-        in_row = self._add_pref_row(in_card, in_row, "Reference images:", "input_ref_dir",
-                                    "Default folder for reference images — Repair Studio and LoRA the Explorer", is_dir=True)
-        in_row = self._add_pref_row(in_card, in_row, "Training images:", "input_dataset_dir",
-                                    "Default folder the Start tab's Browse opens in", is_dir=True)
-
-        self._add_runpod_card(outer)
-
-
-        # Card 4: Actions
-        actions_card = self._start_section_card(outer, "Actions", None)
-        action_row = tk.Frame(actions_card, bg=COLORS["bg_surface"])
-        action_row.pack(anchor=tk.W)
-        ttk.Button(action_row, text="Reset to Defaults", command=self._reset_prefs).pack(side=tk.LEFT, padx=(0, 12))
-        ttk.Button(action_row, text="Open prefs.json", command=self._open_prefs_file).pack(side=tk.LEFT)
-
-        self._add_youtube_help_button(outer, "preferences")
-
-    # Flip to True only when the RunPod template is public AND the URL below is the real one.
-    # Until then the desktop card says "coming soon" instead of offering a button that goes
-    # nowhere — so this can ship long before the template does. test_runpod_card.py refuses to
-    # let LIVE be True while the URL is still the placeholder.
-    RUNPOD_TEMPLATE_LIVE = True
-    RUNPOD_GUIDE_URL = ("https://github.com/shootthesound/Fizgig/blob/master/docker/README.md")
-    # Pre-selects an RTX 5090: it is the cheapest card that clears Fizgig's 32 GB
-    # no-block-swap threshold for Krea 2, so the default lands people on the fastest
-    # sensible option rather than the biggest or the cheapest.
-    RUNPOD_DEPLOY_URL = ("https://console.runpod.io/deploy"
-                         "?type=GPU&gpu=RTX+5090&count=1&template=faoq8ed6um")
-    RUNPOD_REFERRAL = "vkb387ep"
-
-    def _runpod_deploy_url(self) -> str:
-        u = self.RUNPOD_DEPLOY_URL
-        return f"{u}&ref={self.RUNPOD_REFERRAL}" if self.RUNPOD_REFERRAL else u
-
-    def _add_runpod_card(self, outer):
-        """One card, two audiences.
-
-        On a pod it is the control panel for things only a rented machine has — an hourly bill, a
-        volume that fills up, and a browser tab people assume is holding the run up. On the desktop
-        it is how someone finds out Fizgig runs on rented hardware at all."""
-        if _running_on_pod():
-            self._build_pod_controls(outer)
-        else:
-            self._build_pod_advert(outer)
-
-    def _build_pod_controls(self, outer):
-        card = self._start_section_card(
-            outer, "RunPod",
-            "Fizgig is running on a rented GPU. These settings only appear here.")
-        self._runpod_card = card
-
-        # The money one. A finished run on an idle rented GPU bills until someone notices.
-        ttk.Checkbutton(
-            card,
-            text="Stop this pod when a training run finishes",
-            variable=self.prefs_vars["runpod_stop_when_done"], onvalue="1", offvalue="0",
-            style="Surface.TCheckbutton").pack(anchor=tk.W)
-        tk.Label(card,
-                 text="Only after a run completes on its own — never after a Pause, a Stop, or a "
-                      "failure, since those are exactly the times you want the machine alive. You "
-                      "get a two-minute countdown you can cancel. This stops the pod, it never "
-                      "terminates it, so your files are still here when you start it again.",
-                 font=(FONT_FAMILY, 9, "italic"), fg=COLORS["text_explain"],
-                 bg=COLORS["bg_surface"], wraplength=760,
-                 justify=tk.LEFT).pack(anchor=tk.W, pady=(2, 2))
-        # The key field. Better here than as a template variable: a public template hands its
-        # variables to everyone who deploys it, so nobody can safely ship a key in one.
-        key_row = tk.Frame(card, bg=COLORS["bg_surface"])
-        key_row.pack(anchor=tk.W, fill=tk.X, pady=(2, 0))
-        tk.Label(key_row, text="RunPod API key:", font=(FONT_FAMILY, 10),
-                 fg=COLORS["text_secondary"], bg=COLORS["bg_surface"]).pack(side=tk.LEFT,
-                                                                            padx=(0, 8))
-        self._pod_key_entry = ttk.Entry(key_row, textvariable=self.prefs_vars["runpod_api_key"],
-                                        width=44, show="•")
-        self._pod_key_entry.pack(side=tk.LEFT)
-        ttk.Button(key_row, text="Clear", width=7,
-                   command=lambda: self.prefs_vars["runpod_api_key"].set("")).pack(side=tk.LEFT,
-                                                                                    padx=(6, 0))
-        self._pod_key_status = tk.Label(key_row, text="", font=(FONT_FAMILY, 9),
-                                        bg=COLORS["bg_surface"])
-        self._pod_key_status.pack(side=tk.LEFT, padx=(10, 0))
-
-        tk.Label(card,
-                 text="Make one at RunPod → Settings → API Keys. The key RunPod gives a pod "
-                      "automatically is pod-scoped and cannot stop pods, which is a RunPod "
-                      "limitation rather than a Fizgig one. Saved to prefs.json on your volume, so "
-                      "it persists across pods — and stays out of any shared template.",
-                 font=(FONT_FAMILY, 9), fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                 wraplength=760, justify=tk.LEFT).pack(anchor=tk.W, pady=(4, 12))
-
-        self.prefs_vars["runpod_api_key"].trace_add(
-            "write", lambda *a: self._refresh_pod_key_status())
-        self._refresh_pod_key_status()
-
-        # Storage — the thing that silently ends a run at 3am.
-        self._pod_storage_lbl = tk.Label(card, text="", font=(FONT_FAMILY, 10),
-                                         fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                                         justify=tk.LEFT, anchor="w")
-        self._pod_storage_lbl.pack(anchor=tk.W, fill=tk.X)
-        self._refresh_pod_storage()
-
-        tk.Label(card,
-                 text="Your files: datasets in /workspace/datasets, models in /workspace/models, "
-                      "finished LoRAs in /workspace/output_loras. Everything under /workspace "
-                      "survives stopping and restarting the pod — anything outside it does not. "
-                      "On the default Volume Disk it goes when you TERMINATE the pod, so stop "
-                      "rather than terminate between sessions; a Network Volume survives that too. "
-                      "Drag files in and out with the file manager on port 8080.",
-                 font=(FONT_FAMILY, 9), fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                 wraplength=760, justify=tk.LEFT).pack(anchor=tk.W, pady=(10, 0))
-
-        # Asked constantly by anyone new to a remote desktop, and the answer is reassuring.
-        tk.Label(card,
-                 text="Closing this browser tab does not stop training. Fizgig runs on the pod, "
-                      "not in your browser — shut the tab, come back later, and the run is still "
-                      "going.",
-                 font=(FONT_FAMILY, 10, "bold"), fg=COLORS["success"], bg=COLORS["bg_surface"],
-                 wraplength=760, justify=tk.LEFT).pack(anchor=tk.W, pady=(10, 0))
-
-        # Image version AND app commit: the template pins the image while the app updates itself
-        # from git at every boot, so they diverge by design and "what are you running?" needs both.
-        bits = []
-        _img = (os.environ.get("FIZGIG_IMAGE_VERSION") or "").strip()
-        if _img:
-            bits.append(f"image {_img}")
-        _sha = _app_commit()
-        if _sha:
-            bits.append(f"app {_sha}")
-        _pid = _pod_id()
-        if _pid:
-            bits.append(f"pod {_pid}")
-        if bits:
-            _gpu = (os.environ.get("RUNPOD_GPU_NAME") or "").replace("+", " ").strip()
-            line = "  ·  ".join(bits) + (f"  ·  {_gpu}" if _gpu else "")
-            # Readable tier despite being a footer: this is the one line a user is asked to
-            # transcribe into a bug report, so 8pt at 2.54:1 was exactly backwards.
-            tk.Label(card, text=line, font=(FONT_FAMILY, 9), fg=COLORS["text_explain"],
-                     bg=COLORS["bg_surface"], wraplength=760,
-                     justify=tk.LEFT).pack(anchor=tk.W, pady=(10, 0))
-
-    def _build_pod_advert(self, outer):
-        card = self._start_section_card(
-            outer, "Run Fizgig on a rented GPU",
-            "Train on whatever card you like, with as much VRAM as you want, billed by the hour. "
-            "Nothing to install, and your own machine stays free while it trains.")
-        self._runpod_card = card
-        tk.Label(card,
-                 text="Fizgig ships as a ready-made image: the full app in your browser, your "
-                      "models and datasets on persistent storage, and drag-and-drop file transfer. "
-                      "Rent a big card for an afternoon instead of buying one.",
-                 font=(FONT_FAMILY, 10), fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                 wraplength=760, justify=tk.LEFT).pack(anchor=tk.W, pady=(0, 10))
-
-        if self.RUNPOD_TEMPLATE_LIVE:
-            row = tk.Frame(card, bg=COLORS["bg_surface"])
-            row.pack(anchor=tk.W)
-            _btn = tk.Button(row, text="  Deploy on RunPod  ",
-                             font=(FONT_FAMILY, 10, "bold"), fg="#FFFFFF", bg="#673AB7",
-                             activebackground="#5E35B1", activeforeground="#FFFFFF",
-                             relief="flat", bd=0, cursor="hand2", padx=16, pady=6,
-                             command=lambda: self._open_url(self._runpod_deploy_url()))
-            _btn.pack(side=tk.LEFT)
-
-            tk.Label(card,
-                     text="Deploying through this link supports Fizgig's development, at no extra "
-                          "cost to you.",
-                     font=(FONT_FAMILY, 9, "italic"), fg=COLORS["text_explain"],
-                     bg=COLORS["bg_surface"]).pack(anchor=tk.W, pady=(6, 0))
-        else:
-            # Shipped before the template is public. A button that goes nowhere is worse than no
-            # button, and saying so plainly is better than hiding the section and surprising
-            # people with it later.
-            tk.Label(card, text="Coming soon",
-                     font=(FONT_FAMILY, 11, "bold"), fg=COLORS["warning"],
-                     bg=COLORS["bg_surface"]).pack(anchor=tk.W)
-            tk.Label(card,
-                     text="The one-click image is built and being tested. This section will get a "
-                          "Deploy button once it is published — nothing to do in the meantime.",
-                     font=(FONT_FAMILY, 9, "italic"), fg=COLORS["text_explain"],
-                     bg=COLORS["bg_surface"], wraplength=760,
-                     justify=tk.LEFT).pack(anchor=tk.W, pady=(2, 6))
-
-        # The guide, in both states: while it says Coming soon this is the only way to read about
-        # it, and once it is live it is where the storage and cost decisions are explained.
-        _guide = tk.Label(card, text="Read the guide: running Fizgig on a rented GPU",
-                          font=(FONT_FAMILY, 10, "underline"), fg=COLORS["accent_hover"],
-                          bg=COLORS["bg_surface"], cursor="hand2")
-        _guide.pack(anchor=tk.W, pady=(4, 0))
-        _guide.bind("<Button-1>", lambda e: self._open_url(self.RUNPOD_GUIDE_URL))
-        tk.Label(card,
-                 text="When Fizgig is running on a pod, this section turns into its controls — "
-                      "stop the pod automatically when training finishes, check storage, and see "
-                      "where your files live.",
-                 font=(FONT_FAMILY, 9), fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                 wraplength=760, justify=tk.LEFT).pack(anchor=tk.W, pady=(10, 0))
-
-    def _pod_stop_key(self) -> str:
-        """The key to stop this pod with: the one saved in Preferences, else a template env var.
-
-        Prefs first because that is the route that scales — a public template cannot carry anyone's
-        key, since template variables are handed to every container deployed from it."""
-        try:
-            k = self.prefs_vars["runpod_api_key"].get().strip()
-        except Exception:
-            k = ""
-        return k or _pod_stop_key_env()
-
-    def _refresh_pod_key_status(self):
-        lbl = getattr(self, "_pod_key_status", None)
-        if lbl is None or not lbl.winfo_exists():
-            return
-        if self._pod_stop_key():
-            lbl.config(text="auto-stop ready", fg=COLORS["success"])
-        else:
-            lbl.config(text="needed for auto-stop", fg=COLORS["warning"])
-
-    def _refresh_pod_storage(self):
-        """Free space on the volume, refreshed while the tab is open."""
-        lbl = getattr(self, "_pod_storage_lbl", None)
-        if lbl is None or not lbl.winfo_exists():
-            return
-        try:
-            import shutil as _sh
-            usage = _sh.disk_usage("/workspace")
-            free_gb, total_gb = usage.free / 1024 ** 3, usage.total / 1024 ** 3
-            colour = COLORS["text_explain"]
-            if total_gb > 10000:
-                # A network volume does not expose its quota to the container — the kernel reports
-                # the host's whole backing pool, so this read said "431035 GB free of 1430281 GB"
-                # for a 100 GB volume. Printing that verbatim looks broken and, worse, implies
-                # there is room when your quota might be nearly full. Say what we actually know.
-                txt = ("Storage: /workspace is a network volume. Its size is set in RunPod and "
-                       "isn't visible from in here — check usage in the RunPod dashboard.")
-            elif total_gb < 60:
-                # This small means the volume never mounted and /workspace is container disk,
-                # which RunPod wipes when the pod stops — a 32 GB model download would evaporate.
-                txt = (f"Storage: only {free_gb:.0f} GB free of {total_gb:.0f} GB — that looks like "
-                       f"container disk, not your volume. Check the template's volume mount path "
-                       f"is /workspace, or your files will vanish when the pod stops.")
-                colour = COLORS["warning"]
-            else:
-                txt = f"Storage: {free_gb:.0f} GB free of {total_gb:.0f} GB on /workspace"
-                if free_gb < 40:
-                    colour = COLORS["warning"]
-            lbl.config(text=txt, fg=colour)
-        except Exception:
-            lbl.config(text="Storage: unavailable")
-        lbl.after(30000, self._refresh_pod_storage)
-
-    def _open_url(self, url):
-        """Open a link. Central so the pod image's browser handling stays in one place."""
-        webbrowser.open(url)
-
-    def _add_fetch_models_row(self, frame, row, family, blurb, optional_label=None):
-        """'Download them all for me' row at the foot of a model-paths card.
-
-        The per-row Download links open a browser and leave you to save the file and paste the
-        path back — five times over, ~32 GB, before Fizgig does anything at all. This does the
-        same job unattended and writes the paths into prefs itself."""
-        bar = tk.Frame(frame, bg=COLORS["bg_surface"])
-        bar.grid(row=row, column=0, columnspan=3, sticky=tk.EW, pady=(14, 2))
-        btn = ttk.Button(bar, text="⬇ Download models for me",
-                         command=lambda f=family: self._start_fetch_models(f))
-        btn.pack(side=tk.LEFT)
-        setattr(self, f"_fetch_btn_{family}", btn)
-        status = tk.Label(bar, text="", font=(FONT_FAMILY, 9),
-                          fg=COLORS["text_secondary"], bg=COLORS["bg_surface"])
-        status.pack(side=tk.LEFT, padx=(12, 0))
-        setattr(self, f"_fetch_status_{family}", status)
-        if optional_label:
-            # Off by default: the only optional MiniMax weight is the 21 GB ref2va DiT, which a
-            # first setup does not need and most runs never use. The point is that the fetcher
-            # CAN get it - before this, its own Download link was the only route.
-            var = tk.BooleanVar(value=False)
-            setattr(self, f"_fetch_optional_{family}", var)
-            ttk.Checkbutton(bar, text=optional_label, variable=var).pack(side=tk.LEFT, padx=(16, 0))
-        tk.Label(frame, text=blurb, font=(FONT_FAMILY, 9, "italic"), fg=COLORS["text_explain"],
-                 bg=COLORS["bg_surface"], wraplength=760, justify=tk.LEFT
-                 ).grid(row=row + 1, column=0, columnspan=3, sticky=tk.W, pady=(0, 2))
-        return row + 2
-
-    _HF_GATED_URLS = [
-        ("Base DiT — accept the licence",
-         "https://huggingface.co/black-forest-labs/FLUX.2-klein-base-9b-fp8"),
-        ("Distilled DiT — accept the licence",
-         "https://huggingface.co/black-forest-labs/FLUX.2-klein-9b-fp8"),
-        ("VAE — accept the licence",
-         "https://huggingface.co/black-forest-labs/FLUX.2-dev"),
-        ("Then create a READ token here",
-         "https://huggingface.co/settings/tokens"),
-    ]
-
-    def _ask_hf_token(self):
-        """Modal token prompt. Returns the token, or '' if cancelled.
-
-        Not simpledialog.askstring: its prompt is a Label, so the four URLs you have to visit
-        would be un-selectable — readable but not copyable, which is useless when the whole
-        point is to go to them. Each URL here is a readonly Entry (selectable, Ctrl+C works)
-        with Copy and Open buttons."""
-        result = {"token": ""}
-        dlg = tk.Toplevel(self.master)
-        dlg.title("HuggingFace token — Klein downloads")
-        dlg.configure(bg=BG_COLOR)
-        dlg.transient(self.master)
-        dlg.resizable(True, False)
-
-        # Buttons packed BOTTOM first so they can never be pushed off the edge (the v2.8.5
-        # caption-dialog fix — same reasoning applies to any dialog with variable content).
-        btn_row = ttk.Frame(dlg)
-        btn_row.pack(side=tk.BOTTOM, pady=(6, 12))
-
-        tk.Label(dlg, text="Black Forest Labs gate the Klein downloads",
-                 font=(FONT_FAMILY, 11, "bold"), fg=COLORS["text_primary"], bg=BG_COLOR
-                 ).pack(anchor=tk.W, padx=14, pady=(14, 2))
-        tk.Label(dlg,
-                 text="They require every user to accept the licence themselves — which is also why "
-                      "Fizgig can't download these for you without a token of your own. Sign in to "
-                      "HuggingFace, accept on all three pages, then paste a read token below. "
-                      "Krea 2 needs none of this; those files aren't gated.",
-                 font=(FONT_FAMILY, 10), fg=COLORS["text_explain"], bg=BG_COLOR,
-                 wraplength=620, justify=tk.LEFT).pack(anchor=tk.W, padx=14, pady=(0, 10))
-
-        rows = tk.Frame(dlg, bg=BG_COLOR)
-        rows.pack(fill=tk.X, padx=14)
-        rows.columnconfigure(0, weight=1)
-
-        def copy(url, btn):
-            self.master.clipboard_clear()
-            self.master.clipboard_append(url)
-            btn.config(text="copied")
-            btn.after(1200, lambda: btn.config(text="Copy"))
-
-        for i, (label, url) in enumerate(self._HF_GATED_URLS):
-            tk.Label(rows, text=label, font=(FONT_FAMILY, 9), fg=COLORS["text_secondary"],
-                     bg=BG_COLOR).grid(row=i * 2, column=0, columnspan=3, sticky=tk.W,
-                                       pady=(6 if i else 0, 0))
-            e = ttk.Entry(rows, width=62)
-            e.insert(0, url)
-            e.config(state="readonly")           # selectable and copyable; not editable
-            e.grid(row=i * 2 + 1, column=0, sticky=tk.EW)
-            cb = ttk.Button(rows, text="Copy", width=8)
-            cb.config(command=lambda u=url, b=cb: copy(u, b))
-            cb.grid(row=i * 2 + 1, column=1, padx=(6, 0))
-            ttk.Button(rows, text="Open", width=8,
-                       command=lambda u=url: webbrowser.open(u)
-                       ).grid(row=i * 2 + 1, column=2, padx=(6, 0))
-
-        ttk.Button(dlg, text="Open all three licence pages",
-                   command=lambda: [webbrowser.open(u) for _, u in self._HF_GATED_URLS[:3]]
-                   ).pack(anchor=tk.W, padx=14, pady=(12, 0))
-
-        tk.Label(dlg, text="Paste your token:", font=(FONT_FAMILY, 10),
-                 fg=COLORS["text_secondary"], bg=BG_COLOR).pack(anchor=tk.W, padx=14, pady=(14, 2))
-        tok = ttk.Entry(dlg, width=62, show="*")
-        tok.pack(anchor=tk.W, padx=14, fill=tk.X)
-        tk.Label(dlg, text="Starts with hf_… Stored only for this download — never written to disk.",
-                 font=(FONT_FAMILY, 8, "italic"), fg=COLORS["text_muted"], bg=BG_COLOR
-                 ).pack(anchor=tk.W, padx=14, pady=(2, 0))
-
-        def ok():
-            result["token"] = tok.get().strip()
-            dlg.destroy()
-
-        ttk.Button(btn_row, text="Cancel", command=dlg.destroy).pack(side=tk.LEFT, padx=5)
-        ttk.Button(btn_row, text="Download", command=ok).pack(side=tk.LEFT, padx=5)
-        tok.bind("<Return>", lambda e: ok())
-
-        dlg.update_idletasks()
-        tok.focus_set()
-        dlg.grab_set()
-        self.master.wait_window(dlg)
-        return result["token"]
-
-    def _start_fetch_models(self, family):
-        """Run the fetcher in a worker thread, streaming progress into the status label."""
-        if getattr(self, "_fetch_running", False):
-            messagebox.showinfo("Already downloading", "A model download is already running.")
-            return
-        token = ""
-        if family == "klein":
-            # Klein's repos are gated: BFL require each user to accept the licence themselves,
-            # which is exactly why these can't be bundled or pre-fetched on anyone's behalf.
-            # An HF_TOKEN already in the environment (the container's documented env var for
-            # exactly this) satisfies the gate with no prompt — only ask when there isn't one.
-            token = os.environ.get("HF_TOKEN", "").strip()
-            if not token:
-                token = self._ask_hf_token()
-                if not token:
-                    return
-
-        btn = getattr(self, f"_fetch_btn_{family}", None)
-        status = getattr(self, f"_fetch_status_{family}", None)
-        if btn:
-            btn.config(state="disabled")
-        self._fetch_running = True
-
-        def worker():
-            import subprocess
-            # Helper models come first and with EVERY family: they're ~1.9 GB against tens of
-            # GB of weights, and whichever button you press you'll hit Florence / the face model
-            # / the translator / Gizmo's Whisper sooner or later. Fetching them up front means
-            # the Captions tab, Look Filter and Transcribe work immediately (and offline), and
-            # they survive abandoning the big download.
-            # Re-running is cheap — everything here is a no-op once present.
-            cmd = [sys.executable, "-m", "fizgig.scripts.fetch_models", "--progress",
-                   "--family", "tools", "--family", family]
-            _opt = getattr(self, f"_fetch_optional_{family}", None)
-            if _opt is not None and _opt.get():
-                cmd.append("--include-optional")
-            env = self._cuda_env_for_subprocess(dict(os.environ))
-            env["PYTHONPATH"] = os.path.join(FIZGIG_DIR, "src")
-            env["PYTHONUNBUFFERED"] = "1"
-            if token.strip():
-                env["HF_TOKEN"] = token.strip()
-            ok = False
-            try:
-                p = subprocess.Popen(cmd, cwd=FIZGIG_DIR, stdout=subprocess.PIPE,
-                                     stderr=subprocess.STDOUT, text=True, env=env,
-                                     encoding="utf-8", errors="replace",
-                                     creationflags=(subprocess.CREATE_NO_WINDOW
-                                                    if os.name == "nt" else 0))
-                self._fetch_proc = p
-                for line in p.stdout:
-                    line = line.rstrip()
-                    if line:
-                        self.master.after(0, lambda l=line: self._on_fetch_progress(family, l))
-                ok = (p.wait() == 0)
-            except Exception as e:
-                self.master.after(0, lambda e=e: self._on_fetch_progress(
-                    family, f"failed: {type(e).__name__}: {e}"))
-            self.master.after(0, lambda: self._on_fetch_done(family, ok))
-
-        self._open_fetch_progress_window(family)
-        threading.Thread(target=worker, daemon=True).start()
-        if status:
-            status.config(text="downloading…", fg=COLORS["text_secondary"])
-
-    def _open_fetch_progress_window(self, family):
-        """Unmissable progress for a multi-GB download.
-
-        A status label beside the button was the only feedback, and hf_hub_download's own
-        progress never reached it (tqdm redraws with carriage returns, so a line-reader sees
-        nothing until the transfer ends). The result was an app that looked frozen for ten
-        minutes and then popped up 'done'."""
-        win = tk.Toplevel(self.master)
-        self._fetch_win = win
-        win.title("Downloading models")
-        win.configure(bg=BG_COLOR)
-        win.transient(self.master)
-        win.protocol("WM_DELETE_WINDOW", lambda: None)   # closing must go through Cancel
-        win.resizable(False, False)
-
-        tk.Label(win, text="Downloading models", font=(FONT_FAMILY, 12, "bold"),
-                 fg=COLORS["text_primary"], bg=BG_COLOR).pack(anchor=tk.W, padx=18, pady=(16, 2))
-        self._fetch_file_lbl = tk.Label(win, text="starting…", font=(FONT_FAMILY, 10),
-                                        fg=COLORS["text_secondary"], bg=BG_COLOR,
-                                        wraplength=460, justify=tk.LEFT, anchor="w")
-        self._fetch_file_lbl.pack(anchor=tk.W, padx=18, pady=(0, 8), fill=tk.X)
-
-        self._fetch_bar = ttk.Progressbar(win, length=460, mode="indeterminate")
-        self._fetch_bar.pack(padx=18)
-        self._fetch_bar.start(12)
-        self._fetch_bar_mode = "indeterminate"
-
-        self._fetch_bytes_lbl = tk.Label(win, text="", font=(FONT_FAMILY, 9),
-                                         fg=COLORS["text_muted"], bg=BG_COLOR, anchor="w")
-        self._fetch_bytes_lbl.pack(anchor=tk.W, padx=18, pady=(6, 0), fill=tk.X)
-        tk.Label(win, text="You can leave this running and keep using other tabs. Interrupted "
-                           "downloads resume where they left off.",
-                 font=(FONT_FAMILY, 9, "italic"), fg=COLORS["text_explain"], bg=BG_COLOR,
-                 wraplength=460, justify=tk.LEFT).pack(anchor=tk.W, padx=18, pady=(8, 0))
-
-        row = ttk.Frame(win)
-        row.pack(pady=(12, 14))
-        ttk.Button(row, text="Cancel", command=self._cancel_fetch_models).pack()
-        win.update_idletasks()
-
-    def _cancel_fetch_models(self):
-        """Stop the download. Safe: partial files are resumable, prefs are only written on success."""
-        p = getattr(self, "_fetch_proc", None)
-        if p is not None and p.poll() is None:
-            try:
-                p.terminate()
-            except Exception:
-                pass
-        if getattr(self, "_fetch_file_lbl", None) is not None:
-            try:
-                self._fetch_file_lbl.config(text="cancelling…")
-            except Exception:
-                pass
-
-    @staticmethod
-    def _fmt_gb(n):
-        return f"{n / 1024 ** 3:.2f} GB"
-
-    def _on_fetch_progress(self, family, line):
-        stripped = line.strip()
-        if stripped.startswith("[progress]"):
-            try:
-                _tag, done, total, name = stripped.split(None, 3)
-                done, total = int(done), int(total)
-            except (ValueError, IndexError):
-                return
-            bar = getattr(self, "_fetch_bar", None)
-            if bar is not None and bar.winfo_exists():
-                if total > 0:
-                    if self._fetch_bar_mode != "determinate":
-                        bar.stop()
-                        bar.config(mode="determinate", maximum=100)
-                        self._fetch_bar_mode = "determinate"
-                    bar["value"] = max(0, min(100, done * 100 / total))
-                self._fetch_file_lbl.config(text=name)
-                self._fetch_bytes_lbl.config(
-                    text=f"{self._fmt_gb(done)} of {self._fmt_gb(total)}"
-                         f"   ({done * 100 // total if total else 0}%)")
-            return
-
-        # Non-progress lines: the [get]/[ok]/[keep]/[gated] narration.
-        status = getattr(self, f"_fetch_status_{family}", None)
-        if status:
-            status.config(text=stripped[:110])
-        lbl = getattr(self, "_fetch_file_lbl", None)
-        if lbl is not None and lbl.winfo_exists() and stripped:
-            lbl.config(text=stripped)
-            bar = getattr(self, "_fetch_bar", None)
-            # A new item that isn't a download (a cache warm, a skip) — back to the pulsing bar
-            # rather than leaving the previous file's percentage sitting there looking stalled.
-            if bar is not None and bar.winfo_exists() and self._fetch_bar_mode == "determinate" \
-                    and not stripped.startswith("[progress]"):
-                bar.config(mode="indeterminate")
-                bar.start(12)
-                self._fetch_bar_mode = "indeterminate"
-                self._fetch_bytes_lbl.config(text="")
-        self.update_console(f"[models] {line}\n")
-
-    def _on_fetch_done(self, family, ok):
-        self._fetch_running = False
-        self._fetch_proc = None
-        win = getattr(self, "_fetch_win", None)
-        if win is not None:
-            try:
-                win.destroy()
-            except Exception:
-                pass
-            self._fetch_win = None
-        btn = getattr(self, f"_fetch_btn_{family}", None)
-        if btn:
-            btn.config(state="normal")
-        # The fetcher wrote prefs.json from another process, so re-read it — otherwise the
-        # entry boxes keep showing the paths from before the download and saving from this
-        # window would write the stale ones straight back over the new ones.
-        try:
-            fresh = load_prefs()
-            for key, var in self.prefs_vars.items():
-                if key in fresh and str(fresh[key]).strip():
-                    var.set(fresh[key])
-            self.prefs.update(fresh)
-        except Exception:
-            pass
-        status = getattr(self, f"_fetch_status_{family}", None)
-        if status:
-            status.config(text="done — paths filled in" if ok else "finished with items missing (see console)",
-                          fg=COLORS["success"] if ok else COLORS["warning"])
-        if ok:
-            messagebox.showinfo("Models ready",
-                                "Downloaded and wired into Preferences.\n\n"
-                                "The paths above are filled in — nothing else to set up.")
-
-    def _add_pref_row(self, frame, row, label, pref_key, hint, is_dir=False, download_url=None, download_note=None,
-                      download_label="Download", download_url2=None, download_label2="Download", download_note2=None):
-        """Add a labeled pref entry with Browse button and hint text. Returns next row index.
-
-        If download_url is set, a download link (text=download_label) is added next to Browse that opens
-        the URL in the user's default browser. A second link (download_url2 / download_label2) can be
-        added too — used to offer e.g. both fp8 and bf16 variants. download_note / download_note2 are
-        appended to the hint line.
-        """
-        ttk.Label(frame, text=label).grid(row=row, column=0, sticky=tk.W, padx=(0, 10), pady=4)
-        ttk.Entry(frame, textvariable=self.prefs_vars[pref_key], width=60).grid(row=row, column=1, sticky=tk.EW, pady=4)
-
-        btn_frame = tk.Frame(frame, bg=COLORS["bg_surface"])
-        btn_frame.grid(row=row, column=2, sticky=tk.W, padx=(8, 0))
-        browse_cmd = (lambda: self._browse_pref_dir(pref_key)) if is_dir else (lambda: self._browse_pref_file(pref_key))
-        ttk.Button(btn_frame, text="Browse", command=browse_cmd).pack(side=tk.LEFT)
-        for _url, _lbl in ((download_url, download_label), (download_url2, download_label2)):
-            if not _url:
-                continue
-            dl_link = tk.Label(btn_frame, text=_lbl,
-                               fg=COLORS["accent_hover"], cursor="hand2",
-                               font=(FONT_FAMILY, 9, "underline"),
-                               bg=COLORS["bg_surface"])
-            dl_link.pack(side=tk.LEFT, padx=(8, 0))
-            dl_link.bind("<Button-1>", lambda e, url=_url: webbrowser.open(url))
-            ToolTip(dl_link, f"Open download page in browser:\n{_url}")
-        row += 1
-        hint_text = hint
-        _notes = [n for n in (download_note, download_note2) if n]
-        if _notes:
-            hint_text = hint + "  ·  " + "  ·  ".join(_notes)
-        tk.Label(frame, text=hint_text,
-                 font=(FONT_FAMILY, 10, "italic"),
-                 fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                 wraplength=760, justify=tk.LEFT).grid(
-            row=row, column=1, columnspan=2, sticky=tk.W, pady=(0, 6)
-        )
-        row += 1
-        return row
-
-    def _browse_pref_file(self, pref_key):
-        filepath = filedialog.askopenfilename(
-            title="Select file",
-            filetypes=[("SafeTensors", "*.safetensors"), ("All files", "*.*")]
-        )
-        if filepath:
-            self.prefs_vars[pref_key].set(filepath)
-
-    def _browse_pref_dir(self, pref_key):
-        dirpath = filedialog.askdirectory(title="Select directory")
-        if dirpath:
-            self.prefs_vars[pref_key].set(dirpath)
-
-    def _reset_prefs(self):
-        if messagebox.askyesno("Reset Preferences", "Restore all paths to defaults?"):
-            for key, default in DEFAULT_PREFS.items():
-                if key in self.prefs_vars:
-                    self.prefs_vars[key].set(default)
-
-    def _open_prefs_file(self):
-        if os.path.exists(PREFS_FILE):
-            self._open_in_file_manager(PREFS_FILE)
-        else:
-            messagebox.showinfo("Info", "prefs.json doesn't exist yet — change a path to create it.")
 
     # endregion
